@@ -9,8 +9,8 @@ import android.location.LocationManager;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Looper;
-import android.support.annotation.NonNull;
-import android.support.v4.content.PermissionChecker;
+import androidx.annotation.NonNull;
+import androidx.core.content.PermissionChecker;
 import android.util.Log;
 
 import com.google.android.gms.common.ConnectionResult;
@@ -36,7 +36,6 @@ import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
-
 public class PluginLocationService extends CordovaPlugin {
   private Activity activity;
   private final String TAG = "PluginLocationService";
