@@ -1,5 +1,7 @@
 package plugin.google.maps;
 
+import com.google.android.libraries.maps.GoogleMap;
+
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaArgs;
 import org.apache.cordova.CordovaInterface;
@@ -9,7 +11,7 @@ import org.json.JSONException;
 
 public interface MyPluginInterface {
   public void initialize(CordovaInterface cordova, CordovaWebView webView);
-  public void setPluginMap(PluginMap pluginMap);
+  public void setMap(GoogleMap map);
   public boolean execute(String action, String rawArgs, CallbackContext callbackContext) throws JSONException;
   public boolean execute(String action, JSONArray args, CallbackContext callbackContext) throws JSONException;
   public boolean execute(String action, CordovaArgs args, CallbackContext callbackContext) throws JSONException;
