@@ -1,6 +1,6 @@
 package plugin.google.maps;
 
-import android.util.LruCache;
+import androidx.collection.LruCache;
 
 import java.util.HashSet;
 
