@@ -15,13 +15,15 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 
 import za.co.twyst.tbxml.TBXML;
 
 
-public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
+public class PluginKmlOverlay extends MyPlugin implements IOverlayPlugin {
   private HashMap<String, Bundle> styles = new HashMap<String, Bundle>();
+  private String wwwDirName = "www";
 
   private enum KML_TAG {
     NOT_SUPPORTED,
@@ -34,6 +36,11 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
     coordinates
   }
 
+  @Override
+  public void setPluginMap(PluginMap map) {
+    // stub
+  }
+
   /**
    * Create kml overlay
    *
@@ -43,11 +50,20 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
    */
   public void create(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
 
-    final JSONObject opts = args.getJSONObject(1);
-    self = this;
+    final JSONObject opts = args.getJSONObject(2);
     if (!opts.has("url")) {
       callbackContext.error("No kml file is specified");
       return;
+    }
+
+
+    try {
+      if (Arrays.asList(cordova.getContext().getAssets().list("")).contains("capacitor.config.json")) {
+        // Capacitor
+        this.wwwDirName = "public";
+      }
+    } catch (Exception e) {
+      // ignore
     }
 
     cordova.getActivity().runOnUiThread(new Runnable() {
@@ -70,6 +86,7 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
         if (!urlStr.contains("://") &&
             !urlStr.startsWith("/") &&
             !urlStr.startsWith("www/") &&
+            !urlStr.startsWith("public/") &&
             !urlStr.startsWith("data:image") &&
             !urlStr.startsWith("./") &&
             !urlStr.startsWith("../")) {
@@ -79,10 +96,10 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
         if (currentPageUrl.startsWith("http://localhost") ||
             currentPageUrl.startsWith("http://127.0.0.1")) {
           if (urlStr.contains("://")) {
-            urlStr = urlStr.replaceAll("http://.+?/", "file:///android_asset/www/");
+            urlStr = urlStr.replaceAll("http://.+?/", "file:///android_asset/" + wwwDirName + "/");
           } else {
             // Avoid WebViewLocalServer (because can not make a connection for some reason)
-            urlStr = "file:///android_asset/www/".concat(urlStr);
+            urlStr = "file:///android_asset/" + wwwDirName + "/";
           }
         }
 
@@ -98,7 +115,7 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
 
         // Avoid WebViewLocalServer (because can not make a connection for some reason)
         if (urlStr.contains("http://localhost") || urlStr.contains("http://127.0.0.1")) {
-          urlStr = urlStr.replaceAll("^http://[^\\/]+\\//", "file:///android_asset/www/");
+          urlStr = urlStr.replaceAll("^http://[^\\/]+\\//", "file:///android_asset/" + wwwDirName);
         }
 
 
@@ -112,6 +129,11 @@ public class PluginKmlOverlay extends MyPlugin implements MyPluginInterface {
         });
       }
     });
+  }
+
+  @Override
+  public void remove(JSONArray args, CallbackContext callbackContext) throws JSONException {
+    // stub
   }
 
   private Bundle loadKml(String urlStr) {
