@@ -1,35 +1,53 @@
 package plugin.google.maps;
 
 import android.content.res.AssetManager;
+import android.util.Log;
 import android.view.View;
 import android.webkit.WebView;
 
-import com.google.android.gms.maps.model.TileOverlay;
-import com.google.android.gms.maps.model.TileOverlayOptions;
+import com.google.android.libraries.maps.model.TileOverlay;
+import com.google.android.libraries.maps.model.TileOverlayOptions;
 
 import org.apache.cordova.CallbackContext;
-import org.apache.cordova.CordovaPreferences;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class PluginTileOverlay extends MyPlugin implements MyPluginInterface {
+import java.util.concurrent.ConcurrentHashMap;
 
+public class PluginTileOverlay extends MyPlugin implements IOverlayPlugin {
+
+  private PluginMap pluginMap;
+  public final ConcurrentHashMap<String, MetaTileOverlay> objects = new ConcurrentHashMap<String, MetaTileOverlay>();
+
+  @Override
+  public void setPluginMap(PluginMap pluginMap) {
+    this.pluginMap = pluginMap;
+  }
+
+  public PluginMap getMapInstance(String mapId) {
+    return (PluginMap) CordovaGoogleMaps.viewPlugins.get(mapId);
+  }
+  public PluginTileOverlay getInstance(String mapId) {
+    PluginMap mapInstance = getMapInstance(mapId);
+    return (PluginTileOverlay) mapInstance.plugins.get(String.format("%s-tileoverlay", mapId));
+  }
 
   /**
    * Create tile overlay
-   *
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
   public void create(final JSONArray args,
     final CallbackContext callbackContext) throws JSONException {
 
 
-    final JSONObject opts = args.getJSONObject(1);
-    final String id = args.getString(2);
+    final JSONObject opts = args.getJSONObject(2);
+    final String id = args.getString(3);
     final int tileSize = opts.getInt("tileSize");
+
+    String tileOverlayId = "tileoverlay_" + id;
+    final MetaTileOverlay meta = new MetaTileOverlay(tileOverlayId);
+    meta.properties = opts;
+    objects.put(tileOverlayId, meta);
 
 
     final TileOverlayOptions options = new TileOverlayOptions();
@@ -49,7 +67,7 @@ public class PluginTileOverlay extends MyPlugin implements MyPluginInterface {
     }
     final boolean _isDebug = isDebug;
 
-    cordova.getActivity().runOnUiThread(new Runnable() {
+    activity.runOnUiThread(new Runnable() {
       @Override
       public void run() {
 
@@ -62,54 +80,35 @@ public class PluginTileOverlay extends MyPlugin implements MyPluginInterface {
           }
         }
         if ("".equals(userAgent)) {
-          View browserView = webView.getView();
-          String browserViewName = browserView.getClass().getName();
-          if ("org.xwalk.core.XWalkView".equals(browserViewName) ||
-              "org.crosswalk.engine.XWalkCordovaView".equals(browserViewName)) {
-
-            CordovaPreferences preferences = webView.getPreferences();
-            // Set xwalk webview settings by Cordova preferences.
-            String xwalkUserAgent = preferences == null ? "" : preferences.getString("xwalkUserAgent", "");
-            if (!xwalkUserAgent.isEmpty()) {
-              userAgent = xwalkUserAgent;
-            }
-
-            String appendUserAgent = preferences.getString("AppendUserAgent", "");
-            if (!appendUserAgent.isEmpty()) {
-              userAgent = userAgent + " " + appendUserAgent;
-            }
-          } else {
-            userAgent = ((WebView) webView.getEngine().getView()).getSettings().getUserAgentString();
-          }
+          userAgent = ((WebView) webView.getEngine().getView()).getSettings().getUserAgentString();
         }
 
         String currentPageUrl = webView.getUrl();
 
-        AssetManager assetManager = cordova.getActivity().getAssets();
+        AssetManager assetManager = activity.getAssets();
         final PluginTileProvider tileProvider = new PluginTileProvider(pluginMap.getOverlayId(), id, webView, assetManager, currentPageUrl, userAgent, tileSize, _isDebug);
-        tileProvider.setOnCacheClear(new PluginTileProvider.OnCacheClear() {
-          @Override
-          public void onCacheClear(int hashCode) {
-            TileOverlay tileOverlay = (TileOverlay)PluginTileOverlay.this.getTileOverlay(hashCode+"");
-            if (tileOverlay != null) {
-              tileOverlay.clearTileCache();
-              System.gc();
-            }
-          }
-        });
+//        tileProvider.setOnCacheClear(new PluginTileProvider.OnCacheClear() {
+//          @Override
+//          public void onCacheClear(int hashCode) {
+//            TileOverlay tileOverlay = (TileOverlay)PluginTileOverlay.this.objects.
+//            if (tileOverlay != null) {
+//              tileOverlay.clearTileCache();
+//              System.gc();
+//            }
+//          }
+//        });
         options.tileProvider(tileProvider);
 
 
-        TileOverlay tileOverlay = map.addTileOverlay(options);
-        //String id = tileOverlay.getId();
+        TileOverlay tileOverlay = pluginMap.getGoogleMap().addTileOverlay(options);
 
-        pluginMap.objects.put("tileoverlay_" + id, tileOverlay);
-        pluginMap.objects.put("tileprovider_" + id, tileProvider);
+        meta.tileOverlay = tileOverlay;
+        meta.tileProvider = tileProvider;
 
         try {
           JSONObject result = new JSONObject();
           result.put("hashCode", id);
-          result.put("__pgmId", "tileoverlay_" + id);
+          result.put("__pgmId", meta.getId());
           callbackContext.success(result);
         } catch (JSONException e) {
           e.printStackTrace();
@@ -119,106 +118,92 @@ public class PluginTileOverlay extends MyPlugin implements MyPluginInterface {
     });
   }
 
-  @SuppressWarnings("unused")
+  @PgmPluginMethod
   public void onGetTileUrlFromJS(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-    String id = args.getString(0);
-    String urlKey = args.getString(1);
-    String tileUrl = args.getString(2);
-    String pluginId = "tileprovider_" + id;
-    if (pluginMap.objects.containsKey(pluginId)) {
-      ((PluginTileProvider)(this.pluginMap.objects.get(pluginId))).onGetTileUrlFromJS(urlKey, tileUrl);
-    }
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.get(tileOverlayId);
+
+    String urlKey = args.getString(2);
+    String tileUrl = args.getString(3);
+
+    Log.d(TAG, "tileOverlayId = " + tileOverlayId + ", meta = " + meta );
+    meta.tileProvider.onGetTileUrlFromJS(urlKey, tileUrl);
     callbackContext.success();
   }
 
   /**
    * set z-index
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
-  @SuppressWarnings("unused")
+  @PgmPluginMethod(runOnUiThread = true)
   public void setZIndex(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-    String id = args.getString(0);
-    float zIndex = (float) args.getDouble(1);
-    this.setFloat("setZIndex", id, zIndex, callbackContext);
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.get(tileOverlayId);
+
+    float zIndex = (float) args.getDouble(2);
+    meta.tileOverlay.setZIndex(zIndex);
+    callbackContext.success();
   }
 
   /**
    * Set visibility for the object
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
+  @PgmPluginMethod(runOnUiThread = true)
   public void setVisible(JSONArray args, CallbackContext callbackContext) throws JSONException {
-    boolean visible = args.getBoolean(1);
-    String id = args.getString(0);
-    this.setBoolean("setVisible", id, visible, callbackContext);
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.get(tileOverlayId);
+
+    boolean visible = args.getBoolean(2);
+    meta.tileOverlay.setVisible(visible);
+    callbackContext.success();
   }
   /**
    * Remove this tile layer
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
+  @PgmPluginMethod(runOnUiThread = true)
   public void remove(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-    String id = args.getString(0);
-    TileOverlay tileOverlay = (TileOverlay)pluginMap.objects.get(id);
-    if (tileOverlay == null) {
-      callbackContext.success();
-      return;
-    }
-    cordova.getActivity().runOnUiThread(new Runnable() {
-      @Override
-      public void run() {
-
-        try {
-          String id = args.getString(0);
-          TileOverlay tileOverlay = (TileOverlay)pluginMap.objects.remove(id);
-          if (tileOverlay != null) {
-            tileOverlay.clearTileCache();
-            tileOverlay.remove();
-            tileOverlay = null;
-          }
-          id = id.replace("tileoverlay_", "tileprovider_");
-
-          if (pluginMap.objects.containsKey(id)) {
-            ((PluginTileProvider)(pluginMap.objects.get(id))).remove();
-          }
-          //pluginMap.objects.put(id, null);
-          PluginTileProvider provider = (PluginTileProvider) pluginMap.objects.remove(id);
-          provider = null;
-          callbackContext.success();
-        } catch (JSONException e) {
-          e.printStackTrace();
-          callbackContext.error("" + e.getMessage());
-
-        }
-      }
-    });
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.remove(tileOverlayId);
+    meta.tileOverlay.remove();
+    meta.tileOverlay = null;
+    meta.tileProvider.remove();;
+    meta.tileProvider = null;
   }
 
   /**
    * Set fadeIn for the object
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
+  @PgmPluginMethod(runOnUiThread = true)
   public void setFadeIn(JSONArray args, CallbackContext callbackContext) throws JSONException {
-    boolean visible = args.getBoolean(1);
-    String id = args.getString(0);
-    this.setBoolean("setFadeIn", id, visible, callbackContext);
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.get(tileOverlayId);
+
+    boolean isFadedIn = args.getBoolean(2);
+    meta.tileOverlay.setFadeIn(isFadedIn);
+    callbackContext.success();
   }
   /**
    * Set opacity for the tile layer
-   * @param args
-   * @param callbackContext
-   * @throws JSONException
    */
+  @PgmPluginMethod(runOnUiThread = true)
   public void setOpacity(JSONArray args, CallbackContext callbackContext) throws JSONException {
+    String mapId = args.getString(0);
+    PluginTileOverlay instance = getInstance(mapId);
+    String tileOverlayId = args.getString(1);
+    MetaTileOverlay meta = instance.objects.get(tileOverlayId);
+
     double opacity = 1 - args.getDouble(1);
-    String id = args.getString(0);
-    this.setFloat("setTransparency", id, (float)opacity, callbackContext);
+    meta.tileOverlay.setTransparency((float)opacity);
+    callbackContext.success();
   }
 
 }
