@@ -1,16 +1,23 @@
 
-
 var BaseArrayClass = require('cordova-plugin-googlemaps.BaseArrayClass');
 
 var geocoder = null;
 var lastRequestTime = 0;
 var QUEUE = new BaseArrayClass();
 QUEUE.on('insert_at', function() {
+  if (!window.google || !window.google.maps) {
+    return;
+  }
   if (QUEUE.getLength() === 1) {
     this.trigger('next');
   }
 });
 QUEUE.one('insert_at', function() {
+
+  if (!window.google || !window.google.maps) {
+    setTimeout(arguments.callee.bind(this), 100);
+    return;
+  }
   geocoder = new google.maps.Geocoder();
   this.trigger('next');
 });
@@ -32,9 +39,11 @@ QUEUE.on('next', function() {
   geocoder.geocode(cmd.geocoderRequest, function(results, status) {
     switch(status) {
     case google.maps.GeocoderStatus.ERROR:
+      self._executing = false;
       cmd.onError('[geocoding] Cannot connect to Google servers');
       return;
     case google.maps.GeocoderStatus.INVALID_REQUEST:
+      self._executing = false;
       cmd.onError('[geocoding] Invalid request for geocoder');
       return;
     case google.maps.GeocoderStatus.OVER_QUERY_LIMIT:
@@ -46,9 +55,11 @@ QUEUE.on('next', function() {
       }, 3000 + Math.floor(Math.random() * 200));
       return;
     case google.maps.GeocoderStatus.REQUEST_DENIED:
+      self._executing = false;
       cmd.onError('[geocoding] Google denited your geocoding request.');
       return;
     case google.maps.GeocoderStatus.UNKNOWN_ERROR:
+      self._executing = false;
       cmd.onError('[geocoding] There was an unknown error. Please try again.');
       return;
     }
