@@ -1,6 +1,5 @@
 
 
-
 var utils = require('cordova/utils'),
   event = require('cordova-plugin-googlemaps.event'),
   BaseClass = require('cordova-plugin-googlemaps.BaseClass'),
@@ -42,24 +41,59 @@ function displayGrayMap(container) {
 function PluginMap(mapId, options) {
   var self = this;
   BaseClass.apply(this);
-  var mapDiv = document.querySelector('[__pluginMapId=\'' + mapId + '\']');
-  mapDiv.style.backgroundColor = 'rgb(229, 227, 223)';
+  var eles = Array.from(document.querySelectorAll('*'));
+  eles = eles.filter(function(e) {
+    return e.__pluginMapId === mapId;
+  });
+  var actualMapDiv = null;
+  if (eles.length === 1) {
+    actualMapDiv = eles[0];
+  }
 
   var container = document.createElement('div');
   container.style.userSelect='none';
   container.style['-webkit-user-select']='none';
   container.style['-moz-user-select']='none';
   container.style['-ms-user-select']='none';
-  mapDiv.style.position = 'relative';
   container.style.position = 'absolute';
   container.style.top = 0;
   container.style.bottom = 0;
   container.style.right = 0;
   container.style.left = 0;
-  mapDiv.insertBefore(container, mapDiv.firstElementChild);
+  actualMapDiv.insertBefore(container, actualMapDiv.firstElementChild);
+
+  var shadowRoot = container.attachShadow({mode: 'open'});
+
+  var style = document.createElement('style');
+  style.setAttribute('type', 'text/css');
+  style.innerHTML = [
+    ':host {color: black;}',
+    'button.gm-control-active>img {display: none;}',
+    'button.gm-control-active>img:nth-child(1) {display: inline;}'
+  ].join("\n");
+  shadowRoot.appendChild(style);
+
+  var mapDiv = document.createElement('div');
+  mapDiv.style.position = 'relative';
+  mapDiv.style.left = 0;
+  mapDiv.style.top = 0;
+  mapDiv.style.right = 0;
+  mapDiv.style.bottom = 0;
+  mapDiv.style.width = "100%";
+  mapDiv.style.height = "100%";
+  shadowRoot.appendChild(mapDiv);
+
+  mapDiv.style.backgroundColor = 'rgb(229, 227, 223)';
+
+
+
+
+  // mapDiv.insertBefore(container, mapDiv.firstElementChild);
+  // shadowRoot.appendChild(container);
 
   self.set('isGoogleReady', false);
   self.set('container', container);
+  // self.set('shadowRoot', shadowRoot);
   self.PLUGINS = {};
 
   Object.defineProperty(self, '__pgmId', {
@@ -97,7 +131,8 @@ function PluginMap(mapId, options) {
       minZoom: 2,
       disableDefaultUI: true,
       zoomControl: true,
-      center: {lat: 0, lng: 0}
+      center: {lat: 0, lng: 0},
+      clickableIcons: true
     };
 
     if (options) {
@@ -113,74 +148,49 @@ function PluginMap(mapId, options) {
           mapInitOptions.zoomControl = options.controls.zoom == true;
         }
       }
+
+      if (options.gestures) {
+        mapInitOptions.draggable = options.gestures.scroll;
+        mapInitOptions.gestureHandling = options.gestures.scroll;
+        mapInitOptions.disableDoubleClickZoom = !options.gestures.zoom;
+      }
       if (options.preferences) {
+
+        if (options.preferences && options.preferences.restriction) {
+          mapInitOptions.restriction = {
+            latLngBounds: {
+              south: options.preferences.restriction.south,
+              west: options.preferences.restriction.west,
+              north: options.preferences.restriction.north,
+              east: options.preferences.restriction.east
+            },
+            strictBounds: false
+          };
+        }
+
         if (options.preferences.zoom) {
           mapInitOptions.minZoom = options.preferences.zoom.minZoom;
           if (options.preferences.zoom.maxZoom) {
             mapInitOptions.maxZoom = options.preferences.zoom.maxZoom;
           }
         }
+
+        if ('clickableIcons' in options.preferences) {
+          mapInitOptions.clickableIcons = options.preferences.clickableIcons === true;
+        }
       }
     }
 
-    var map = new google.maps.Map(container, mapInitOptions);
+    var map = new google.maps.Map(mapDiv, mapInitOptions);
     map.mapTypes = mapTypeReg;
     self.set('map', map);
 
-    var boundsLimit = null;
-    if (options.preferences && options.preferences.gestureBounds &&
-        options.preferences.gestureBounds.length > 0) {
-      boundsLimit = new google.maps.LatLngBounds();
-      options.preferences.gestureBounds.forEach(function(pos) {
-        boundsLimit.extend(pos);
-      });
-    }
-    map.set('boundsLimit', boundsLimit);
 
     var timeoutError = setTimeout(function() {
       self.trigger('load_error');
       displayGrayMap(mapDiv);
     }, 3000);
 
-    map.addListener('bounds_changed', function() {
-      var boundsLimit = map.get('boundsLimit');
-      if (!boundsLimit) {
-        return;
-      }
-      var visibleBounds = map.getBounds();
-      if (boundsLimit.intersects(visibleBounds) ||
-          visibleBounds.contains(boundsLimit.getNorthEast()) && visibleBounds.contains(boundsLimit.getSouthWest()) ||
-          boundsLimit.contains(visibleBounds.getNorthEast()) && boundsLimit.contains(visibleBounds.getSouthWest())) {
-        return;
-      }
-      var center = map.getCenter();
-      var dummyLat = center.lat(),
-        dummyLng = center.lng();
-      var ne = boundsLimit.getNorthEast(),
-        sw = boundsLimit.getSouthWest();
-      if (dummyLat < sw.lat() ) {
-        dummyLat = sw.lat();
-      } else if (dummyLat > ne.lat()) {
-        dummyLat = ne.lat();
-      }
-      if (dummyLng < 0) {
-        // the Western Hemisphere
-        if (dummyLng > ne.lng()) {
-          dummyLng = ne.lng();
-        } else if (dummyLng < sw.lng()) {
-          dummyLng = sw.lng();
-        }
-      } else {
-        // the Eastern Hemisphere
-        if (dummyLng > ne.lng()) {
-          dummyLng = ne.lng();
-        } else if (dummyLng < sw.lng()) {
-          dummyLng = sw.lng();
-        }
-      }
-      var dummyLatLng = new google.maps.LatLng(dummyLat, dummyLng);
-      map.panTo(dummyLatLng);
-    });
 
     google.maps.event.addListenerOnce(map, 'projection_changed', function() {
       clearTimeout(timeoutError);
@@ -224,7 +234,7 @@ function PluginMap(mapId, options) {
             options.camera.target.forEach(function(pos) {
               bounds.extend(pos);
             });
-            map.fitBounds(bounds, 5);
+            map.fitBounds(bounds, 'padding' in options.camera ? options.camera.padding || 0 : 5);
           } else {
             map.setCenter(options.camera.target);
           }
@@ -251,12 +261,45 @@ function PluginMap(mapId, options) {
 
 utils.extend(PluginMap, BaseClass);
 
+PluginMap.prototype._cmd = function(onSuccess, onError, args) {
+
+  var self = this;
+  var info = args[0];
+  if (!info.instance) {
+    return onError(new Error('info.instance is missing'));
+  }
+  // if (!(info.instance in self.objects)) {
+  //   console.log(self.objects);
+  //   return onError(new Error(info.instance + 'is not found.'));
+  // }
+
+  var tmp = info.instance.split('_');
+  var className;
+  if (info.instance.indexOf('-tileoverlay') > -1) {
+    className = 'tileoverlay';
+  } else {
+    className = (info.instance.split("_"))[0];
+  }
+  var pluginClass = self.PLUGINS[className];
+  if (!pluginClass) {
+    return onError(new Error('Invalid instance id "' + info.instance + '".'));
+  }
+
+  pluginClass[info.cmd].call(pluginClass, onSuccess, onError, info.args);
+};
+
 PluginMap.prototype.setOptions = function(onSuccess, onError, args) {
   var self = this;
   var map = self.get('map'),
     options = args[0];
 
-  var mapInitOptions = {};
+  var mapInitOptions = {
+    draggable: true,
+    gestureHandling: 'auto',
+    disableDoubleClickZoom: false,
+    heading: 0,
+    tilt: 0
+  };
 
   if (options) {
     if (options.mapType) {
@@ -268,26 +311,51 @@ PluginMap.prototype.setOptions = function(onSuccess, onError, args) {
 
     if (options.controls) {
       if (options.controls.zoom !== undefined) {
-        mapInitOptions.zoomControl = options.controls.zoom == true;
+        mapInitOptions.zoomControl = options.controls.zoom === true;
       }
     }
+    if (options.gestures) {
+      mapInitOptions.draggable = options.gestures.scroll === true;
+      mapInitOptions.gestureHandling = options.gestures.scroll === true;
+      mapInitOptions.disableDoubleClickZoom = !(options.gestures.zoom === true);
+    }
+
     if (options.preferences) {
-      if (options.preferences.zoom) {
-        mapInitOptions.minZoom = Math.max(options.preferences.zoom || 2, 2);
-        if (options.preferences.zoom.maxZoom) {
-          mapInitOptions.maxZoom = options.preferences.zoom.maxZoom;
+      if ('zoom' in options.preferences) {
+        if (options.preferences.zoom) {
+          if ('minZoom' in options.preferences.zoom) {
+            mapInitOptions.minZoom = Math.max(options.preferences.zoom.minZoom, 2);
+          } else {
+            mapInitOptions.minZoom = undefined;
+          }
+          if ('maxZoom' in options.preferences.zoom) {
+            mapInitOptions.maxZoom = Math.min(options.preferences.zoom.maxZoom, 23);
+          } else {
+            mapInitOptions.maxZoom = undefined;
+          }
+        } else {
+          mapInitOptions.zoom = undefined;
         }
       }
 
-      if ('gestureBounds' in options.preferences) {
-        var boundsLimit = null;
-        if (options.preferences.gestureBounds && options.preferences.gestureBounds.length > 0) {
-          boundsLimit = new google.maps.LatLngBounds();
-          options.preferences.gestureBounds.forEach(function(pos) {
-            boundsLimit.extend(pos);
-          });
+      if ('restriction' in options.preferences) {
+        if (options.preferences.restriction) {
+          mapInitOptions.restriction = {
+            latLngBounds: {
+              south: options.preferences.restriction.south,
+              west: options.preferences.restriction.west,
+              north: options.preferences.restriction.north,
+              east: options.preferences.restriction.east
+            },
+            strictBounds: false
+          };
+        } else {
+          mapInitOptions.restriction = undefined;
         }
-        map.set('boundsLimit', boundsLimit);
+      }
+
+      if ('clickableIcons' in options.preferences) {
+        mapInitOptions.clickableIcons = options.preferences.clickableIcons === true;
       }
 
     }
@@ -327,6 +395,7 @@ PluginMap.prototype.setOptions = function(onSuccess, onError, args) {
   onSuccess();
 };
 
+
 PluginMap.prototype.setActiveMarkerId = function(onSuccess, onError, args) {
   var self = this,
     markerId = args[0];
@@ -347,22 +416,44 @@ PluginMap.prototype.setDiv = function(onSuccess, onError, args) {
   var self = this,
     map = self.get('map'),
     container = self.get('container');
-
-  if (args.length === 0) {
-    if (container && container.parentNode) {
-      container.parentNode.removeAttribute('__pluginMapId');
-      container.parentNode.removeChild(container);
-    }
-  } else {
-    var domId = args[0];
-    var mapDiv = document.querySelector('[__pluginDomId=\'' + domId + '\']');
-    mapDiv.style.position = 'relative';
-    mapDiv.insertBefore(container, mapDiv.firstElementChild);
-    mapDiv.setAttribute('__pluginMapId', self.__pgmId);
+    // shadowRoot = self.get('shadowRoot');
+  if (!container) {
+    onError('[map.setDiv] container is undefined');
+    return;
   }
 
-  google.maps.event.trigger(map, 'resize');
-  onSuccess();
+  if (args.length === 0) {
+    if (container.parentNode && container.parentNode.parentNode) {
+      container.parentNode.parentNode.removeChild(container.parentNode);
+    }
+    onSuccess();
+  } else {
+
+    if (container.parentNode && !container.parentNode.parentNode) {
+      var domId = args[0];
+      var eles = Array.from(document.querySelectorAll('*'));
+      eles = eles.filter(function(e) {
+        return e.__pluginDomId === domId;
+      });
+      var actualMapDiv = eles[0];
+      Object.defineProperty(actualMapDiv, '__pluginMapId', {
+        enumerable: false,
+        value: self.__pgmId
+      });
+      actualMapDiv.style.position = 'relative';
+      actualMapDiv.insertBefore(container, actualMapDiv.firstElementChild);
+
+      var offsetHeight = container.offsetHeight;
+      container.offsetHeight = offsetHeight - 1;
+      setTimeout(function() {
+        container.offsetHeight = offsetHeight;
+        google.maps.event.trigger(map, 'resize');
+      }, 1000);
+    }
+
+    onSuccess();
+  }
+
 };
 PluginMap.prototype.resizeMap = function(onSuccess) {
   var self = this;
@@ -696,28 +787,29 @@ PluginMap.prototype._onCameraEvent = function(evtName) {
 PluginMap.prototype.loadPlugin = function(onSuccess, onError, args) {
   var self = this;
   var className = args[0];
+  var lowerClassName = className.toLowerCase();
 
   var plugin;
-  if (className in self.PLUGINS) {
-    plugin = self.PLUGINS[className];
+  if (lowerClassName in self.PLUGINS) {
+    plugin = self.PLUGINS[lowerClassName];
   } else {
     var OverlayClass = require('cordova-plugin-googlemaps.Plugin' + className);
     plugin = new OverlayClass(this);
-    self.PLUGINS[className] = plugin;
+    self.PLUGINS[lowerClassName] = plugin;
 
-    // Since Cordova involes methods as Window,
-    // the `this` keyword of involved method is Window, not overlay itself.
-    // In order to keep indicate the `this` keyword as overlay itself,
-    // wrap the method.
-    var dummyObj = {};
-    for (var key in OverlayClass.prototype) {
-      if (typeof OverlayClass.prototype[key] === 'function') {
-        dummyObj[key] = plugin[key].bind(plugin);
-      } else {
-        dummyObj[key] = plugin[key];
-      }
-    }
-    require('cordova/exec/proxy').add(self.__pgmId + '-' + className.toLowerCase(), dummyObj);
+    // // Since Cordova involes methods as Window,
+    // // the `this` keyword of involved method is Window, not overlay itself.
+    // // In order to keep indicate the `this` keyword as overlay itself,
+    // // wrap the method.
+    // var dummyObj = {};
+    // for (var key in OverlayClass.prototype) {
+    //   if (typeof OverlayClass.prototype[key] === 'function') {
+    //     dummyObj[key] = plugin[key].bind(plugin);
+    //   } else {
+    //     dummyObj[key] = plugin[key];
+    //   }
+    // }
+    // require('cordova/exec/proxy').add(self.__pgmId + '-' + className.toLowerCase(), dummyObj);
   }
 
   plugin._create.call(plugin, onSuccess, onError, args);
