@@ -13,16 +13,20 @@
 #import "PluginUtil.h"
 #import "PluginMap.h"
 #import "PluginStreetViewPanorama.h"
-#import "MyPluginLayer.h"
+#import "PgmPluginLayer.h"
 #import "IPluginProtocol.h"
-#import "IPluginView.h"
+
+static NSMutableDictionary *viewPlugins;
 
 @interface CordovaGoogleMaps : CDVPlugin
 
-@property (nonatomic) MyPluginLayer *pluginLayer;
-@property (nonatomic) NSMutableDictionary *viewPlugins;
+@property (nonatomic) BOOL isSdkAvailable;
+@property (nonatomic) PgmPluginLayer *pluginLayer;
 @property (nonatomic) NSOperationQueue *executeQueue;
 
++ (id)getViewPlugin:(NSString *)pluginId;
+
+- (void)updateMapPositionOnly:(CDVInvokedUrlCommand *)command;
 - (void)getMap:(CDVInvokedUrlCommand*)command;
 - (void)getPanorama:(CDVInvokedUrlCommand*)command;
 - (void)clearHtmlElements:(CDVInvokedUrlCommand *)command;
@@ -30,4 +34,11 @@
 - (void)removeMap:(CDVInvokedUrlCommand *)command;
 - (void)pause:(CDVInvokedUrlCommand *)command;
 - (void)resume:(CDVInvokedUrlCommand *)command;
+- (void)setDiv:(CDVInvokedUrlCommand *)command;
+- (void)attachToWebView:(CDVInvokedUrlCommand*)command;
+- (void)detachFromWebView:(CDVInvokedUrlCommand*)command;
+- (void)resizeMap:(CDVInvokedUrlCommand *)command;
+- (void)setBackGroundColor:(CDVInvokedUrlCommand *)command;
+- (void)getLicenseInfo:(CDVInvokedUrlCommand*)command;
+- (void)setEnv:(CDVInvokedUrlCommand*)command;
 @end
