@@ -22,33 +22,29 @@
 
 - (void)pluginUnload
 {
-
-
-    // Plugin destroy
-    NSArray *keys = [self.mapCtrl.objects allKeys];
-    NSString *key;
-    for (int i = 0; i < [keys count]; i++) {
-        key = [keys objectAtIndex:i];
-        if ([key hasPrefix:@"groundoverlay_property"]) {
-          key = [key stringByReplacingOccurrencesOfString:@"_property" withString:@""];
-          GMSGroundOverlay *groundoverlay = (GMSGroundOverlay *)[self.mapCtrl.objects objectForKey:key];
-          groundoverlay.map = nil;
-          groundoverlay = nil;
-        }
-        [self.mapCtrl.objects removeObjectForKey:key];
+  
+  // Plugin destroy
+  NSArray *keys = [self.mapCtrl.objects allKeys];
+  NSString *key;
+  for (int i = 0; i < [keys count]; i++) {
+    key = [keys objectAtIndex:i];
+    if ([key hasPrefix:@"groundoverlay_property"]) {
+      key = [key stringByReplacingOccurrencesOfString:@"_property" withString:@""];
+      GMSGroundOverlay *groundoverlay = (GMSGroundOverlay *)[self.mapCtrl.objects objectForKey:key];
+      if (groundoverlay != nil) {
+        groundoverlay.map = nil;
+      }
+      groundoverlay = nil;
     }
-
-  //[self.imgCache removeAllObjects];
-  //self.imgCache = nil;
+    [self.mapCtrl.objects removeObjectForKey:key];
+  }
 
   key = nil;
   keys = nil;
 
+
   NSString *pluginId = [NSString stringWithFormat:@"%@-groundoverlay", self.mapCtrl.overlayId];
-  CDVViewController *cdvViewController = (CDVViewController*)self.viewController;
-  [cdvViewController.pluginObjects removeObjectForKey:pluginId];
-  [cdvViewController.pluginsMap setValue:nil forKey:pluginId];
-  pluginId = nil;
+  [self.mapCtrl.plugins removeObjectForKey:pluginId];
 }
 
 -(void)setPluginViewController:(PluginViewController *)viewCtrl
@@ -56,12 +52,18 @@
     self.mapCtrl = (PluginMapViewController *)viewCtrl;
 }
 
+-(PluginGroundOverlay *)_getInstance: (NSString *)mapId {
+  NSString *pluginId = [NSString stringWithFormat:@"%@-groundoverlay", mapId];
+  PluginMap *mapInstance = [CordovaGoogleMaps getViewPlugin:mapId];
+  return [mapInstance.mapCtrl.plugins objectForKey:pluginId];
+}
+
 -(void)create:(CDVInvokedUrlCommand *)command
 {
     PluginGroundOverlay *self_ = self;
 
-    NSDictionary *json = [command.arguments objectAtIndex:1];
-    NSString *idBase = [command.arguments objectAtIndex:2];
+    NSDictionary *json = [command.arguments objectAtIndex:2];
+    NSString *idBase = [command.arguments objectAtIndex:3];
     NSArray *points = [json objectForKey:@"bounds"];
 
     GMSMutablePath *path = [GMSMutablePath path];
@@ -78,103 +80,101 @@
     }
     bounds = [[GMSCoordinateBounds alloc] initWithPath:path];
 
-    dispatch_async(dispatch_get_main_queue(), ^{
-        GMSGroundOverlay *groundOverlay = [GMSGroundOverlay groundOverlayWithBounds:bounds icon:nil];
+    GMSGroundOverlay *groundOverlay = [GMSGroundOverlay groundOverlayWithBounds:bounds icon:nil];
 
-        NSString *groundOverlayId = [NSString stringWithFormat:@"groundoverlay_%@", idBase];
-        [self.mapCtrl.objects setObject:groundOverlay forKey: groundOverlayId];
-        groundOverlay.title = groundOverlayId;
-        groundOverlay.anchor = CGPointMake(0.5f, 0.5f);
+    NSString *groundOverlayId = [NSString stringWithFormat:@"groundoverlay_%@", idBase];
+    [self.mapCtrl.objects setObject:groundOverlay forKey: groundOverlayId];
+    groundOverlay.title = groundOverlayId;
+    groundOverlay.anchor = CGPointMake(0.5f, 0.5f);
 
-        if ([json valueForKey:@"zIndex"] && [json valueForKey:@"zIndex"] != [NSNull null]) {
-            groundOverlay.zIndex = [[json valueForKey:@"zIndex"] floatValue];
-        }
+    if ([json valueForKey:@"zIndex"] && [json valueForKey:@"zIndex"] != [NSNull null]) {
+        groundOverlay.zIndex = [[json valueForKey:@"zIndex"] floatValue];
+    }
 
-        if ([json valueForKey:@"bearing"] && [json valueForKey:@"bearing"] != [NSNull null]) {
-            groundOverlay.bearing = [[json valueForKey:@"bearing"] floatValue];
-        }
-        if ([json valueForKey:@"anchor"] && [json valueForKey:@"anchor"] != [NSNull null]) {
-            NSArray *anchor = [json valueForKey:@"anchor"];
-            groundOverlay.anchor = CGPointMake([[anchor objectAtIndex:0] floatValue], [[anchor objectAtIndex:1] floatValue]);
-        }
+    if ([json valueForKey:@"bearing"] && [json valueForKey:@"bearing"] != [NSNull null]) {
+        groundOverlay.bearing = [[json valueForKey:@"bearing"] floatValue];
+    }
+    if ([json valueForKey:@"anchor"] && [json valueForKey:@"anchor"] != [NSNull null]) {
+        NSArray *anchor = [json valueForKey:@"anchor"];
+        groundOverlay.anchor = CGPointMake([[anchor objectAtIndex:0] floatValue], [[anchor objectAtIndex:1] floatValue]);
+    }
 
-        BOOL isVisible = YES;
+    BOOL isVisible = YES;
 
-        // Visible property
-        NSString *visibleValue = [NSString stringWithFormat:@"%@",  json[@"visible"]];
-        if ([@"0" isEqualToString:visibleValue]) {
-          // false
-          isVisible = NO;
-          groundOverlay.map = nil;
-        } else {
-          // true or default
-          groundOverlay.map = self.mapCtrl.map;
-        }
-        BOOL isClickable = NO;
-        if ([json valueForKey:@"clickable"] != [NSNull null] && [[json valueForKey:@"clickable"] boolValue]) {
-            isClickable = YES;
-        }
-
-
-        // Since this plugin uses own touch-detection,
-        // set NO to the tappable property.
-        groundOverlay.tappable = NO;
-
-        __block PluginGroundOverlay *me = self;
-
-        // Load image
-        [self.mapCtrl.executeQueue addOperationWithBlock:^{
-            NSString *urlStr = [json objectForKey:@"url"];
-            if (urlStr) {
-                [self _setImage:groundOverlay urlStr:urlStr completionHandler:^(BOOL successed) {
-                  if (!successed) {
-                      CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR];
-                      [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-                      return;
-                  }
-
-                  //NSString *imgId = [NSString stringWithFormat:@"groundoverlay_image_%lu", (unsigned long)groundOverlay.hash];
-                  //[me.imgCache setObject:groundOverlay.icon forKey:imgId];
-
-                  if ([json valueForKey:@"opacity"] && [json valueForKey:@"opacity"] != [NSNull null]) {
-                      CGFloat opacity = [[json valueForKey:@"opacity"] floatValue];
-                      groundOverlay.icon = [groundOverlay.icon imageByApplyingAlpha:opacity];
-                  }
+    // Visible property
+    NSString *visibleValue = [NSString stringWithFormat:@"%@",  json[@"visible"]];
+    if ([@"0" isEqualToString:visibleValue]) {
+      // false
+      isVisible = NO;
+      groundOverlay.map = nil;
+    } else {
+      // true or default
+      groundOverlay.map = self.mapCtrl.map;
+    }
+    BOOL isClickable = NO;
+    if ([json valueForKey:@"clickable"] != [NSNull null] && [[json valueForKey:@"clickable"] boolValue]) {
+        isClickable = YES;
+    }
 
 
-                  //---------------------------
-                  // Keep the properties
-                  //---------------------------
-                  NSString *propertyId = [NSString stringWithFormat:@"groundoverlay_property_%@", idBase];
+    // Since this plugin uses own touch-detection,
+    // set NO to the tappable property.
+    groundOverlay.tappable = NO;
 
-                  // points
-                  NSMutableDictionary *properties = [[NSMutableDictionary alloc] init];
-                  // bounds (pre-calculate for click detection)
-                  [properties setObject:groundOverlay.bounds  forKey:@"bounds"];
-                  // isVisible
-                  [properties setObject:[NSNumber numberWithBool:isVisible] forKey:@"isVisible"];
-                  // isClickable
-                  [properties setObject:[NSNumber numberWithBool:isClickable] forKey:@"isClickable"];
-                  // zIndex
-                  [properties setObject:[NSNumber numberWithFloat:groundOverlay.zIndex] forKey:@"zIndex"];;
-                  [me.mapCtrl.objects setObject:properties forKey:propertyId];
+    __block PluginGroundOverlay *me = self;
 
-                  //---------------------------
-                  // Result for JS
-                  //---------------------------
-                  NSMutableDictionary *result = [[NSMutableDictionary alloc] init];
-                  [result setObject:groundOverlayId forKey:@"__pgmId"];
+    // Load image
+    [self.mapCtrl.executeQueue addOperationWithBlock:^{
+      NSString *urlStr = [json objectForKey:@"url"];
+      if (urlStr) {
+        [self _setImage:groundOverlay urlStr:urlStr completionHandler:^(BOOL successed) {
+          if (!successed) {
+            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR];
+            [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+            return;
+          }
 
-                  CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result];
-                  [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+          //NSString *imgId = [NSString stringWithFormat:@"groundoverlay_image_%lu", (unsigned long)groundOverlay.hash];
+          //[me.imgCache setObject:groundOverlay.icon forKey:imgId];
 
-              }];
-            } else {
-                CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"Error: The url property is not specified."];
-                [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-            }
+          if ([json valueForKey:@"opacity"] && [json valueForKey:@"opacity"] != [NSNull null]) {
+            CGFloat opacity = [[json valueForKey:@"opacity"] floatValue];
+            groundOverlay.icon = [groundOverlay.icon imageByApplyingAlpha:opacity];
+          }
+
+
+          //---------------------------
+          // Keep the properties
+          //---------------------------
+          NSString *propertyId = [NSString stringWithFormat:@"groundoverlay_property_%@", idBase];
+
+          // points
+          NSMutableDictionary *properties = [[NSMutableDictionary alloc] init];
+          // bounds (pre-calculate for click detection)
+          [properties setObject:groundOverlay.bounds  forKey:@"bounds"];
+          // isVisible
+          [properties setObject:[NSNumber numberWithBool:isVisible] forKey:@"isVisible"];
+          // isClickable
+          [properties setObject:[NSNumber numberWithBool:isClickable] forKey:@"isClickable"];
+          // zIndex
+          [properties setObject:[NSNumber numberWithFloat:groundOverlay.zIndex] forKey:@"zIndex"];;
+          [me.mapCtrl.objects setObject:properties forKey:propertyId];
+
+          //---------------------------
+          // Result for JS
+          //---------------------------
+          NSMutableDictionary *result = [[NSMutableDictionary alloc] init];
+          [result setObject:groundOverlayId forKey:@"__pgmId"];
+
+          CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result];
+          [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+
         }];
-    });
+      } else {
+        CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"Error: The url property is not specified."];
+        [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+      }
+    }];
 
 }
 
@@ -222,10 +222,8 @@
                 range = [urlStr rangeOfString:@"/"];
                 if (range.location != 0) {
                   // Get the current URL, then calculate the relative path.
-                  CDVViewController *cdvViewController = (CDVViewController*)self.viewController;
-                  id webview = cdvViewController.webView;
                   dispatch_sync(dispatch_get_main_queue(), ^{
-                       NSURL *url = [webview URL];
+                    NSURL *url = [(WKWebView *)self.webView URL];
                        NSString *currentURL = url.absoluteString;
                        if (![[url lastPathComponent] isEqualToString:@"/"]) {
                          currentURL = [currentURL stringByDeletingLastPathComponent];
@@ -245,7 +243,7 @@
                        urlStr2 = [urlStr2 stringByReplacingOccurrencesOfString:@":///" withString:@"://"];
                        url = [NSURL URLWithString:urlStr2];
 
-                       [self downloadImageWithURL:url  completionBlock:^(BOOL succeeded, UIImage *image) {
+                       [PluginUtil downloadImageWithURL:url  completionBlock:^(BOOL succeeded, UIImage *image) {
 
                          if (!succeeded) {
                            completionHandler(NO);
@@ -280,323 +278,263 @@
             image = [UIImage imageNamed:urlStr];
         }
         dispatch_async(dispatch_get_main_queue(), ^{
-            groundOverlay.icon = [UIImage imageNamed:urlStr];
-            completionHandler(YES);
+          groundOverlay.icon = [UIImage imageNamed:urlStr];
+          completionHandler(YES);
         });
 
 
     } else {
-        NSURL *url = [NSURL URLWithString:urlStr];
-        [self downloadImageWithURL:url  completionBlock:^(BOOL succeeded, UIImage *image) {
+      NSURL *url = [NSURL URLWithString:urlStr];
+      [PluginUtil downloadImageWithURL:url  completionBlock:^(BOOL succeeded, UIImage *image) {
 
-            if (!succeeded) {
-              completionHandler(NO);
-              return;
-            }
+        if (!succeeded) {
+          completionHandler(NO);
+          return;
+        }
 
-            dispatch_async(dispatch_get_main_queue(), ^{
-                groundOverlay.icon = image;
-                completionHandler(YES);
-            });
+        dispatch_async(dispatch_get_main_queue(), ^{
+            groundOverlay.icon = image;
+            completionHandler(YES);
+        });
 
-        }];
+      }];
     }
 }
 
 /**
  * Remove the ground overlay
- * @params key
+ * 
  */
 -(void)remove:(CDVInvokedUrlCommand *)command
 {
-    [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+  GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
 
-        //[self.imgCache removeObjectForKey:groundOverlayId];
+  //[self.imgCache removeObjectForKey:groundOverlayId];
 
-        NSString *propertyId = [groundOverlayId stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property"];
-        [self.mapCtrl.objects removeObjectForKey:propertyId];
-        [self.mapCtrl.objects removeObjectForKey:groundOverlayId];
+  NSString *propertyId = [groundOverlayId stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property"];
+  [goverlayInstance.mapCtrl.objects removeObjectForKey:propertyId];
+  [goverlayInstance.mapCtrl.objects removeObjectForKey:groundOverlayId];
 
-        groundOverlay.map = nil;
-        groundOverlay = nil;
-        [self.mapCtrl.objects removeObjectForKey:groundOverlayId];
+  groundOverlay.map = nil;
+  groundOverlay = nil;
+  [goverlayInstance.mapCtrl.objects removeObjectForKey:groundOverlayId];
 
-        CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-        [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-    }];
+  CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+  [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
 }
 
 
 /**
  * Set visibility
- * @params key
+ * 
  */
 -(void)setVisible:(CDVInvokedUrlCommand *)command
 {
+  
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
 
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
+    NSString *key = [command.arguments objectAtIndex:1];
+    GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:key];
+    Boolean isVisible = [[command.arguments objectAtIndex:2] boolValue];
 
-        NSString *key = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:key];
-        Boolean isVisible = [[command.arguments objectAtIndex:1] boolValue];
+    // Update the property
+    NSString *propertyId = [key stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property_"];
+    NSMutableDictionary *properties = [NSMutableDictionary dictionaryWithDictionary:
+                                       [goverlayInstance.mapCtrl.objects objectForKey:propertyId]];
+    [properties setObject:[NSNumber numberWithBool:isVisible] forKey:@"isVisible"];
+    [goverlayInstance.mapCtrl.objects setObject:properties forKey:propertyId];
 
-        // Update the property
-        NSString *propertyId = [key stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property_"];
-        NSMutableDictionary *properties = [NSMutableDictionary dictionaryWithDictionary:
-                                           [self.mapCtrl.objects objectForKey:propertyId]];
-        [properties setObject:[NSNumber numberWithBool:isVisible] forKey:@"isVisible"];
-        [self.mapCtrl.objects setObject:properties forKey:propertyId];
-
-        [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-            if (isVisible) {
-              groundOverlay.map = self.mapCtrl.map;
-            } else {
-              groundOverlay.map = nil;
-            }
-
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }];
-
+    [[NSOperationQueue mainQueue] addOperationWithBlock:^{
+        if (isVisible) {
+          groundOverlay.map = goverlayInstance.mapCtrl.map;
+        } else {
+          groundOverlay.map = nil;
+        }
     }];
+
+    CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+  }];
 }
 
 /**
  * set image
- * @params key
+ * 
  */
 -(void)setImage:(CDVInvokedUrlCommand *)command
 {
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
 
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
 
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
-        NSString *urlStr = [command.arguments objectAtIndex:1];
-        if (urlStr) {
-            __block PluginGroundOverlay *self_ = self;
-            [self _setImage:groundOverlay urlStr:urlStr completionHandler:^(BOOL successed) {
-                CDVPluginResult* pluginResult;
-                if (successed) {
-                    pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-                } else {
-                    pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR];
-                }
+    NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+    GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
+    NSString *urlStr = [command.arguments objectAtIndex:2];
+    if (urlStr) {
+      __block PluginGroundOverlay *self_ = goverlayInstance;
+      [goverlayInstance _setImage:groundOverlay urlStr:urlStr completionHandler:^(BOOL successed) {
+          CDVPluginResult* pluginResult;
+          if (successed) {
+              pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+          } else {
+              pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR];
+          }
 
-                [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-            }];
-        } else {
+          [self_.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+      }];
+    } else {
 
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }
+      CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+      [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+    }
 
-    }];
+  }];
 
 }
 
 /**
  * Set bounds
- * @params key
+ * 
  */
 -(void)setBounds:(CDVInvokedUrlCommand *)command
 {
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
-        GMSMutablePath *path = [GMSMutablePath path];
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
+      NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+      GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
+      GMSMutablePath *path = [GMSMutablePath path];
 
-        NSArray *points = [command.arguments objectAtIndex:1];
-        int i = 0;
-        NSDictionary *latLng;
-        for (i = 0; i < points.count; i++) {
-            latLng = [points objectAtIndex:i];
-            [path addCoordinate:CLLocationCoordinate2DMake([[latLng objectForKey:@"lat"] floatValue], [[latLng objectForKey:@"lng"] floatValue])];
-        }
-        GMSCoordinateBounds *bounds;
-        bounds = [[GMSCoordinateBounds alloc] initWithPath:path];
+      NSArray *points = [command.arguments objectAtIndex:2];
+      int i = 0;
+      NSDictionary *latLng;
+      for (i = 0; i < points.count; i++) {
+          latLng = [points objectAtIndex:i];
+          [path addCoordinate:CLLocationCoordinate2DMake([[latLng objectForKey:@"lat"] floatValue], [[latLng objectForKey:@"lng"] floatValue])];
+      }
+      GMSCoordinateBounds *bounds;
+      bounds = [[GMSCoordinateBounds alloc] initWithPath:path];
 
-        [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-            [groundOverlay setBounds:bounds];
+      [[NSOperationQueue mainQueue] addOperationWithBlock:^{
+          [groundOverlay setBounds:bounds];
+      }];
 
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }];
-    }];
+      CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+      [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+  }];
 }
 
 /**
  * Set opacity
- * @params key
+ * 
  */
 -(void)setOpacity:(CDVInvokedUrlCommand *)command
 {
+  
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
+    NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+    GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
+    CGFloat opacity = [[command.arguments objectAtIndex:2] floatValue];
 
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
-        CGFloat opacity = [[command.arguments objectAtIndex:1] floatValue];
-
-        [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-            groundOverlay.opacity = opacity;
-
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }];
+    [[NSOperationQueue mainQueue] addOperationWithBlock:^{
+        groundOverlay.opacity = opacity;
     }];
+
+    CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+  }];
 }
 
 /**
  * Set bearing
- * @params key
+ * 
  */
 -(void)setBearing:(CDVInvokedUrlCommand *)command
 {
-
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
-        [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-            groundOverlay.bearing = [[command.arguments objectAtIndex:1] floatValue];
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }];
-
+  
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
+    NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+    GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
+    
+    [[NSOperationQueue mainQueue] addOperationWithBlock:^{
+        groundOverlay.bearing = [[command.arguments objectAtIndex:2] floatValue];
     }];
+    
+    CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+
+  }];
 
 }
 
 
 /**
  * Set clickable
- * @params key
+ * 
  */
 -(void)setClickable:(CDVInvokedUrlCommand *)command
 {
-  [self.mapCtrl.executeQueue addOperationWithBlock:^{
+  
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
 
-      NSString *key = [command.arguments objectAtIndex:0];
-      //GMSGroundOverlay *groundOverlay = (GMSGroundOverlay *)[self.mapCtrl.objects objectForKey:key];
-      Boolean isClickable = [[command.arguments objectAtIndex:1] boolValue];
+    NSString *key = [command.arguments objectAtIndex:1];
+    //GMSGroundOverlay *groundOverlay = (GMSGroundOverlay *)[self.mapCtrl.objects objectForKey:key];
+    Boolean isClickable = [[command.arguments objectAtIndex:2] boolValue];
 
-      // Update the property
-      NSString *propertyId = [key stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property_"];
-      NSMutableDictionary *properties = [NSMutableDictionary dictionaryWithDictionary:
-                                         [self.mapCtrl.objects objectForKey:propertyId]];
-      [properties setObject:[NSNumber numberWithBool:isClickable] forKey:@"isClickable"];
-      [self.mapCtrl.objects setObject:properties forKey:propertyId];
+    // Update the property
+    NSString *propertyId = [key stringByReplacingOccurrencesOfString:@"groundoverlay_" withString:@"groundoverlay_property_"];
+    NSMutableDictionary *properties = [NSMutableDictionary dictionaryWithDictionary:
+                                       [goverlayInstance.mapCtrl.objects objectForKey:propertyId]];
+    [properties setObject:[NSNumber numberWithBool:isClickable] forKey:@"isClickable"];
+    [goverlayInstance.mapCtrl.objects setObject:properties forKey:propertyId];
 
-      CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-      [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+    CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
   }];
 }
 
 /**
  * Set z-index
- * @params key
+ * 
  */
 -(void)setZIndex:(CDVInvokedUrlCommand *)command
 {
+  
+  NSString *mapId = [command.arguments objectAtIndex:0];
+  PluginGroundOverlay *goverlayInstance = [self _getInstance:mapId];
+  
+  [goverlayInstance.mapCtrl.executeQueue addOperationWithBlock:^{
+    NSString *groundOverlayId = [command.arguments objectAtIndex:1];
+    GMSGroundOverlay *groundOverlay = [goverlayInstance.mapCtrl.objects objectForKey:groundOverlayId];
 
-    [self.mapCtrl.executeQueue addOperationWithBlock:^{
-        NSString *groundOverlayId = [command.arguments objectAtIndex:0];
-        GMSGroundOverlay *groundOverlay = [self.mapCtrl.objects objectForKey:groundOverlayId];
-
-        NSInteger zIndex = [[command.arguments objectAtIndex:1] integerValue];
-        [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-            [groundOverlay setZIndex:(int)zIndex];
-
-            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
-            [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-        }];
+    NSInteger zIndex = [[command.arguments objectAtIndex:2] integerValue];
+    [[NSOperationQueue mainQueue] addOperationWithBlock:^{
+        [groundOverlay setZIndex:(int)zIndex];
     }];
-}
 
-
-- (void)downloadImageWithURL:(NSURL *)url completionBlock:(void (^)(BOOL succeeded, UIImage *image))completionBlock
-{
-  [self.mapCtrl.executeQueue addOperationWithBlock:^{
-
-    NSString *iconPath = url.absoluteString;
-
-    // Since ionic local server declines HTTP access for some reason,
-    // replace URL with file path
-    NSBundle *mainBundle = [NSBundle mainBundle];
-    NSString *wwwPath = [mainBundle pathForResource:@"www/cordova" ofType:@"js"];
-    wwwPath = [wwwPath stringByReplacingOccurrencesOfString:@"/cordova.js" withString:@""];
-    if ([iconPath containsString:@"assets/"]) {
-      iconPath = [iconPath regReplace:@"^.*assets/" replaceTxt:[NSString stringWithFormat:@"%@/assets/", wwwPath] options:NSRegularExpressionCaseInsensitive];
-    }
-    // iconPath = [iconPath stringByReplacingOccurrencesOfString:wwwPath withString: @""];
-    
-    // ionic 4
-    iconPath = [iconPath stringByReplacingOccurrencesOfString:@"http://localhost:8080" withString: wwwPath];
-
-    // ionic 5
-    iconPath = [iconPath stringByReplacingOccurrencesOfString:@"ionic://localhost" withString: wwwPath];
-    
-    if ([iconPath hasPrefix:@"file://"] || [iconPath hasPrefix:@"/"]) {
-      iconPath = [iconPath stringByReplacingOccurrencesOfString:@"file://" withString:@""];
-      if (![iconPath hasPrefix:@"/"]) {
-        iconPath = [NSString stringWithFormat:@"/%@", iconPath];
-      }
-      NSFileManager *fileManager = [NSFileManager defaultManager];
-      if (![fileManager fileExistsAtPath:iconPath]) {
-        //if (self.mapCtrl.debuggable) {
-        NSLog(@"(error)There is no file at '%@'.", iconPath);
-        //}
-        //[self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
-        completionBlock(NO, nil);
-        return;
-      }
-
-      UIImage *image = [UIImage imageNamed:iconPath];
-      if (image) {
-        completionBlock(YES, image);
-        return;
-      }
-    }
-
-
-
-    NSURLRequest *req = [NSURLRequest requestWithURL:url
-                                         cachePolicy:NSURLRequestReturnCacheDataElseLoad
-                                     timeoutInterval:5];
-    NSCachedURLResponse *cachedResponse = [[NSURLCache sharedURLCache] cachedResponseForRequest:req];
-    if (cachedResponse != nil) {
-      UIImage *image = [[UIImage alloc] initWithData:cachedResponse.data];
-      if (image) {
-        completionBlock(YES, image);
-        return;
-      }
-    }
-
-
-    //-------------------------------------------------------------
-    // Use NSURLSessionDataTask instead of [NSURLConnection sendAsynchronousRequest]
-    // https://stackoverflow.com/a/20871647
-    //-------------------------------------------------------------
-    NSURLSessionConfiguration *sessionConfiguration = [NSURLSessionConfiguration defaultSessionConfiguration];
-    NSURLSession *session = [NSURLSession sessionWithConfiguration:sessionConfiguration];
-    NSURLSessionDataTask *getTask = [session dataTaskWithRequest:req
-                                               completionHandler:^(NSData *data, NSURLResponse *res, NSError *error) {
-                                                 [session finishTasksAndInvalidate];
-
-                                                 UIImage *image = [UIImage imageWithData:data];
-                                                 if (image) {
-                                                   completionBlock(YES, image);
-                                                   return;
-                                                 }
-
-                                                 completionBlock(NO, nil);
-
-                                               }];
-    [getTask resume];
-
-
+    CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [goverlayInstance.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
   }];
 }
+
 
 @end
