@@ -8,9 +8,9 @@
 
 #import "CordovaGoogleMaps.h"
 #import "IPluginProtocol.h"
-#import "IPluginView.h"
 
-@interface PluginMap : CDVPlugin<IPluginProtocol, IPluginView>
+
+@interface PluginMap : CDVPlugin<IPluginProtocol>
 @property (nonatomic, strong) PluginMapViewController* mapCtrl;
 @property (nonatomic) BOOL isRemoved;
 @property (nonatomic) BOOL initialized;
@@ -22,7 +22,6 @@
 - (void)setCameraTarget:(CDVInvokedUrlCommand*)command;
 - (void)setCameraBearing:(CDVInvokedUrlCommand *)command;
 - (void)setCameraZoom:(CDVInvokedUrlCommand*)command;
-- (void)setDiv:(CDVInvokedUrlCommand *)command;
 - (void)setMapTypeId:(CDVInvokedUrlCommand*)command;
 - (void)animateCamera:(CDVInvokedUrlCommand*)command;
 - (void)loadPlugin:(CDVInvokedUrlCommand*)command;
@@ -32,8 +31,6 @@
 - (void)setIndoorEnabled:(CDVInvokedUrlCommand*)command;
 - (void)setTrafficEnabled:(CDVInvokedUrlCommand*)command;
 - (void)setCompassEnabled:(CDVInvokedUrlCommand*)command;
-- (void)attachToWebView:(CDVInvokedUrlCommand*)command;
-- (void)detachFromWebView:(CDVInvokedUrlCommand*)command;
 - (void)toDataURL:(CDVInvokedUrlCommand*)command;
 - (void)setOptions:(CDVInvokedUrlCommand*)command;
 - (void)setAllGesturesEnabled:(CDVInvokedUrlCommand*)command;
