@@ -17,7 +17,7 @@
 
 @interface PluginStreetViewPanoramaController : PluginViewController<GMSPanoramaViewDelegate>
 
-@property (nonatomic) GMSPanoramaView* panoramaView;
+@property (nonatomic) GMSPanoramaView *panoramaView;
 
 //----------------------------------------------------
 // In order to keep the compatibility with Android,
