@@ -18,23 +18,6 @@ NSDictionary *debugAttributes;
 - (id)initWithOptions:(NSDictionary *) options webView:(UIView *)webView{
     self = [super init];
     self.webView  = webView;
-    //self.tileUrlFormat = [options objectForKey:@"tileUrlFormat"];
-    //NSString *urlStr = [options objectForKey:@"wwwPath"];
-  
-  
-    // Since ionic local server declines HTTP access for some reason,
-    // replace URL with file path
-    NSBundle *mainBundle = [NSBundle mainBundle];
-    NSString *urlStr = [mainBundle pathForResource:@"www/cordova" ofType:@"js"];
-    urlStr = [urlStr stringByReplacingOccurrencesOfString:@"/cordova.js" withString:@""];
-
-    // ionic 4
-    urlStr = [urlStr stringByReplacingOccurrencesOfString:@"http://localhost:8080" withString: urlStr];
-    
-    // ionic 5
-    urlStr = [urlStr stringByReplacingOccurrencesOfString:@"ionic://localhost" withString: urlStr];
-    self.wwwPath = urlStr;
-  
   
     if ([options objectForKey:@"tileSize"]) {
         self.tileSize = [[options objectForKey:@"tileSize"] floatValue];
@@ -64,8 +47,6 @@ NSDictionary *debugAttributes;
       };
     }
 
-    self.imgCache = [[NSCache alloc]init];
-    self.imgCache.totalCostLimit = 3 * 1024 * 1024 * 1024; // 3MB = Cache for image
     self.executeQueue =  [NSOperationQueue new];
 
     return self;
@@ -80,6 +61,7 @@ NSDictionary *debugAttributes;
 }
 
 - (void)onGetTileUrlFromJS:(NSString *)urlKey tileUrl:(NSString *)tileUrl {
+  
     @synchronized (self.tileUrlMap) {
       [self.tileUrlMap setObject:tileUrl forKey:urlKey];
       dispatch_semaphore_signal(self.semaphore);
@@ -148,23 +130,8 @@ NSDictionary *debugAttributes;
     return;
   }
 
-  NSRange range = [urlStr rangeOfString:@"http"];
-  if (range.location == 0) {
-      //-------------------------
-      // http:// or https://
-      //-------------------------
-      [self downloadImageWithX:x y:y zoom:zoom url:[NSURL URLWithString:urlStr] receiver:receiver];
-      return;
-  }
-  urlStr = [urlStr stringByReplacingOccurrencesOfString:@"file://" withString:@""];
-  
-  // ionic 4
-  urlStr = [urlStr stringByReplacingOccurrencesOfString:@"http://localhost:8080" withString: self.wwwPath];
-  
-  // ionic 5
-  urlStr = [urlStr stringByReplacingOccurrencesOfString:@"ionic://localhost" withString: self.wwwPath];
-  
-  if ([urlStr hasPrefix:@"http://"] || [urlStr hasPrefix:@"https://"]) {
+  NSRange range = [urlStr rangeOfString:@"://"];
+  if (range.location != -1) {
       //-------------------------
       // http:// or https://
       //-------------------------
@@ -273,97 +240,22 @@ NSDictionary *debugAttributes;
 
 - (void)downloadImageWithX:(NSUInteger)x y:(NSUInteger)y  zoom:(NSUInteger)zoom  url:(NSURL *)url receiver: (id<GMSTileReceiver>) receiver
 {
-  [self.executeQueue addOperationWithBlock:^{
+  [PluginUtil downloadImageWithURL:url completionBlock:^(BOOL succeeded, UIImage *image) {
 
-    NSURLRequest *req = [NSURLRequest requestWithURL:url
-                                         cachePolicy:NSURLRequestReturnCacheDataElseLoad
-                                     timeoutInterval:5];
-    NSCachedURLResponse *cachedResponse = [[NSURLCache sharedURLCache] cachedResponseForRequest:req];
-    if (cachedResponse != nil) {
-      UIImage *image = [[UIImage alloc] initWithData:cachedResponse.data];
-      if (image) {
-        if (self.isDebug) {
-          image = [self drawDebugInfoWithImage:image
-                                             x:x
-                                             y:y
-                                             zoom:zoom
-                                             url: url.absoluteString];
-        }
-        [receiver receiveTileWithX:x y:y zoom:zoom image:image];
-        return;
+    if (image) {
+      if (self.isDebug) {
+        image = [self drawDebugInfoWithImage:image
+                                            x:x
+                                            y:y
+                                            zoom:zoom
+                                            url: url.absoluteString];
       }
+      [receiver receiveTileWithX:x y:y zoom:zoom image:image];
+      return;
     }
-
-    NSString *uniqueKey = url.absoluteString;
-    NSData *cache = [self.imgCache objectForKey:uniqueKey];
-    if (cache != nil) {
-      UIImage *image = [[UIImage alloc] initWithData:cache];
-      if (image) {
-        if (self.isDebug) {
-          image = [self drawDebugInfoWithImage:image
-                                             x:x
-                                             y:y
-                                             zoom:zoom
-                                             url: url.absoluteString];
-        }
-        [receiver receiveTileWithX:x y:y zoom:zoom image:image];
-        return;
-      }
-    }
-
-    //-------------------------------------------------------------
-    // Use NSURLSessionDataTask instead of [NSURLConnection sendAsynchronousRequest]
-    // https://stackoverflow.com/a/20871647
-    //-------------------------------------------------------------
-    NSURLSessionConfiguration *sessionConfiguration = [NSURLSessionConfiguration defaultSessionConfiguration];
-    NSURLSession *session = [NSURLSession sessionWithConfiguration:sessionConfiguration];
-    NSURLSessionDataTask *getTask = [session dataTaskWithRequest:req
-                             completionHandler:^(NSData *data, NSURLResponse *res, NSError *error) {
-                               [session finishTasksAndInvalidate];
-                               if ( !error ) {
-                                 [self.imgCache setObject:data forKey:uniqueKey cost:data.length];
-                                 UIImage *image = [UIImage imageWithData:data];
-                                 if (self.isDebug) {
-                                   image = [self drawDebugInfoWithImage:image
-                                                                       x:x
-                                                                       y:y
-                                                                       zoom:zoom
-                                                                       url: url.absoluteString];
-                                 }
-                                 [receiver receiveTileWithX:x y:y zoom:zoom image:image];
-                               } else {
-                                 if (self.isDebug) {
-                                   UIImage *image = [self drawDebugInfoWithImage:nil
-                                                                       x:x
-                                                                       y:y
-                                                                       zoom:zoom
-                                                                       url: url.absoluteString];
-                                   [receiver receiveTileWithX:x y:y zoom:zoom image:image];
-                                 } else {
-                                   [receiver receiveTileWithX:x y:y zoom:zoom image:kGMSTileLayerNoTile];
-                                 }
-                               }
-
-                            }];
-    [getTask resume];
-    //-------------------------------------------------------------
-    // NSURLConnection sendAsynchronousRequest is deprecated.
-    //-------------------------------------------------------------
-/*
-    [NSURLConnection sendAsynchronousRequest:req
-                                       queue:self.executeQueue
-                           completionHandler:^(NSURLResponse *res, NSData *data, NSError *error) {
-                             if ( !error ) {
-                               [self.imgCache setObject:data forKey:uniqueKey cost:data.length];
-                               UIImage *image = [UIImage imageWithData:data];
-                               [receiver receiveTileWithX:x y:y zoom:zoom image:image];
-                             } else {
-                               [receiver receiveTileWithX:x y:y zoom:zoom image:kGMSTileLayerNoTile];
-                             }
-
-                           }];
-*/
+    
   }];
+  
 
 }
 
