@@ -6,21 +6,31 @@
 //
 //
 
+#if __has_include("CapacitorCordova.h")
+  #define PGM_PLATFORM_CAPACITOR
+#else
+  #define PGM_PLATFORM_CORDOVA
+#endif
+
+
 #ifndef MIN
-#import <NSObjCRuntime.h>
+  #import <NSObjCRuntime.h>
 #endif
 #import <Foundation/Foundation.h>
 #import <Cordova/CDV.h>
-#import "MainViewController.h"
+
+#ifdef PGM_PLATFORM_CORDOVA
+  #import "MainViewController.h"
+#endif
+
 #import <QuartzCore/QuartzCore.h>
 #import <objc/runtime.h>
-//#import "MFGoogleMapAdditions/GMSCoordinateBounds+Geometry.h"
-#import "GMSCoordinateBounds+Geometry.h"
 #import <math.h>
 #import "IPluginProtocol.h"
 #import "PluginViewController.h"
 #import <Cordova/CDVCommandDelegate.h>
 #import <Cordova/CDVCommandDelegateImpl.h>
+#import "UIImageCache.h"
 
 typedef void (^MYCompletionHandler)(NSError *error);
 
@@ -40,6 +50,7 @@ typedef void (^MYCompletionHandler)(NSError *error);
 @end
 
 @interface NSString (GoogleMapsPlugin)
+- (NSString *)urlencode;
 - (NSString*)regReplace:(NSString*)pattern replaceTxt:(NSString*)replaceTxt options:(NSRegularExpressionOptions)options;
 @end
 
@@ -49,15 +60,11 @@ typedef void (^MYCompletionHandler)(NSError *error);
 - (UIImage *)resize:(CGFloat)width height:(CGFloat)height;
 @end
 
-@interface CDVCommandDelegateImpl (GoogleMapsPlugin)
-- (void)hookSendPluginResult:(CDVPluginResult*)result callbackId:(NSString*)callbackId;
-@end
-
 //
 // animationDidStop for group animation
 // http://stackoverflow.com/a/28051909/697856
 //
-typedef void (^TIFAnimationGroupCompletionBlock)();
+typedef void (^TIFAnimationGroupCompletionBlock)(void);
 @interface CAAnimationGroup (Blocks)
 - (void)setCompletionBlock:(TIFAnimationGroupCompletionBlock)handler;
 @end
@@ -72,6 +79,12 @@ typedef void (^TIFAnimationGroupCompletionBlock)();
 + (GMSMutablePath *)getMutablePathFromCircle:(CLLocationCoordinate2D)center radius:(double)radius;
 + (NSString *)getAbsolutePathFromCDVFilePath:(UIView*)theWebView cdvFilePath:(NSString *)cdvFilePath;
 + (NSString *)PGM_LOCALIZATION:(NSString *)key;
++ (void)getJsonWithURL:(NSString *)baseUrlStr params:(NSDictionary *)params completionBlock:(void (^)(BOOL succeeded, NSDictionary *response, NSString *error))completionBlock;
++ (void)getJsonWithURL:(NSURL *)url completionBlock:(void (^)(BOOL succeeded, NSDictionary *response, NSString *error))completionBlock;
++ (double)getZoomFromBounds:(GMSCoordinateBounds *)bounds mapWidth:(double)mapWidth mapHeight:(double)mapHeight;
++ (double)_zoom:(double)mapPx worldPx:(double)worldPx fraction:(double)fraction;
++ (double)_latRad:(double)lat;
++ (void)downloadImageWithURL:(NSURL *)url completionBlock:(void (^)(BOOL succeeded, UIImage *image))completionBlock;
 @end
 
 
