@@ -1,6 +1,5 @@
 
 var BaseArrayClass = require('./BaseArrayClass');
-var utils = require('cordova/utils');
 
 var resolvedPromise = typeof Promise == 'undefined' ? null : Promise.resolve();
 var nextTick = resolvedPromise ? function (fn) {
@@ -45,11 +44,14 @@ function HTMLColor2RGBA(colorValue, defaultOpacity) {
   }
   if (colorStr.match(/^#([0-9A-F]){3}$/i)) {
     matches = colorStr.match(/([0-9A-F])/ig);
+    var r = parseInt(matches[0], 16);
+    var g = parseInt(matches[1], 16);
+    var b = parseInt(matches[2], 16);
 
     return [
-      parseInt(matches[0], 16),
-      parseInt(matches[1], 16),
-      parseInt(matches[2], 16),
+      (r << 4) + r,
+      (g << 4) + g,
+      (b << 4) + b,
       alpha
     ];
   }
@@ -59,10 +61,14 @@ function HTMLColor2RGBA(colorValue, defaultOpacity) {
     alpha = parseInt(alpha + alpha, 16);
 
     matches = colorStr.match(/([0-9A-F])/ig);
+    var r = parseInt(matches[0], 16);
+    var g = parseInt(matches[1], 16);
+    var b = parseInt(matches[2], 16);
+
     return [
-      parseInt(matches[0], 16),
-      parseInt(matches[1], 16),
-      parseInt(matches[2], 16),
+      (r << 4) + r,
+      (g << 4) + g,
+      (b << 4) + b,
       alpha
     ];
   }
@@ -295,11 +301,11 @@ function getZIndex(dom) {
   if (dom.currentStyle) {
     z = dom.currentStyle['z-index'];
   }
-  var elemId = dom.getAttribute('__pluginDomId');
+  var elemId = dom.__pluginDomId;
   var parentNode = dom.parentNode;
   var parentZIndex = 0;
   if (parentNode && parentNode.nodeType === Node.ELEMENT_NODE) {
-    var parentElemId = parentNode.getAttribute('__pluginDomId');
+    var parentElemId = parentNode.__pluginDomId;
     if (parentElemId in internalCache) {
       parentZIndex = internalCache[parentElemId];
     } else {
@@ -549,8 +555,8 @@ function getLatLng(target) {
 function convertToPositionArray(array) {
   array = array || [];
 
-  if (!utils.isArray(array)) {
-    if (array.type === 'LatLngBounds') {
+  if (!Array.isArray(array)) {
+    if (array.type === 'LatLngBounds' || array.southwest && array.northeast) {
       array = [
         array.southwest,
         {
@@ -677,14 +683,17 @@ function getPluginDomId(element) {
   if (!element || !shouldWatchByNative(element)) {
     return;
   }
-  var elemId = element.getAttribute('__pluginDomId');
+  var elemId = element.__pluginDomId;
   if (!elemId) {
     if (element === document.body) {
       elemId = 'root';
     } else {
       elemId = 'pgm' + Math.floor(Math.random() * Date.now());
     }
-    element.setAttribute('__pluginDomId', elemId);
+    Object.defineProperty(element, '__pluginDomId', {
+      enumerable: false,
+      value: elemId
+    });
   }
   return elemId;
 }
@@ -724,6 +733,11 @@ function createEvent(eventName, properties) {
   return evt;
 }
 
+function hasTransparentClass(div) {
+  return (div.classList && div.classList.contains('_gmaps_cdv_') ||
+    div.className && div.className.indexOf('_gmaps_cdv_') > -1);
+}
+
 function attachTransparentClass(div) {
 
   if (div.classList && !div.classList.contains('_gmaps_cdv_')) {
@@ -734,36 +748,32 @@ function attachTransparentClass(div) {
 
   if (div.shadowRoot) {
     var styleAttr = div.getAttribute('style') || '';
-    if (styleAttr && styleAttr.indexOf('--ion-background-color') === -1) {
-      styleAttr = styleAttr + ' --ion-background-color: transparent;';
+    if (styleAttr && styleAttr.indexOf('--pgm-background-color') === -1) {
+      styleAttr = styleAttr + ' --ion-background-color: var(--pgm-background-color);';
     }
     div.setAttribute('style', styleAttr);
   }
 }
-// function dettachTransparentClass(root) {
-//
-//   if (div.classList && div.classList.contains('_gmaps_cdv_')) {
-//     div.classList.remove('_gmaps_cdv_');
-//   } else if (div.className && div.className.indexOf('_gmaps_cdv_') === -1) {
-//     div.className = div.className.replace('_gmaps_cdv_', '');
-//   }
-//
-//   var visibilityCSS = getStyle(node, 'visibility');
-//   if (getStyle(div, 'background-color') === 'transparent !important') {
-//     div.style.backgroundColor = undefined;
-//   }
-//   if (getStyle(div, 'background') === 'transparent !important') {
-//     div.style.backgroundColor = undefined;
-//   }
-//   if (getStyle(div, 'background-image') === 'transparent !important') {
-//     div.style.backgroundImage = undefined;
-//   }
-//   if (div.shadowRoot) {
-//     var hiddenChildren = div.querySelectorAll('*');
-//     hiddenChildren = Array.prototype.splice(hiddenChildren, 0);
-//     hiddenChildren.forEach(dettachTransparentClass);
-//   }
-// }
+function detachTransparentClass(div) {
+
+  if (div.classList && div.classList.contains('_gmaps_cdv_')) {
+    div.classList.remove('_gmaps_cdv_');
+  } else if (div.className && div.className.indexOf('_gmaps_cdv_') > -1) {
+    div.className = div.className.replace('_gmaps_cdv_', '');
+  }
+
+  if (div.shadowRoot) {
+    var styleAttr = div.getAttribute('style') || '';
+    if (styleAttr && styleAttr.indexOf('--pgm-background-color') > -1) {
+      styleAttr = styleAttr.replace('--ion-background-color: var(--pgm-background-color);', '');
+    }
+    div.setAttribute('style', styleAttr);
+
+    var hiddenChildren = div.querySelectorAll('*');
+    hiddenChildren = Array.prototype.splice(hiddenChildren, 0);
+    hiddenChildren.forEach(detachTransparentClass);
+  }
+}
 
 module.exports = {
   _clearInternalCache: _clearInternalCache,
@@ -787,8 +797,9 @@ module.exports = {
   getPluginDomId: getPluginDomId,
   hashCode: hashCode,
   createEvent: createEvent,
-  //dettachTransparentClass: dettachTransparentClass,
-  attachTransparentClass: attachTransparentClass
+  detachTransparentClass: detachTransparentClass,
+  attachTransparentClass: attachTransparentClass,
+  hasTransparentClass: hasTransparentClass
 };
 
 if (cordova && cordova.platformId === 'browser') {
