@@ -21,31 +21,37 @@ var GroundOverlay = function (map, groundOverlayOptions, _exec) {
   //-----------------------------------------------
   self.on('visible_changed', function () {
     var visible = self.get('visible');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setVisible', [self.getId(), visible]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setVisible', [self.map.getId(), self.getId(), visible]);
   });
   self.on('image_changed', function () {
     var image = self.get('image');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setImage', [self.getId(), image]);
+
+    var link = document.createElement('a');
+    link.href = image;
+    image = link.protocol+'//'+link.host+link.pathname + link.search;
+    link = undefined;
+
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setImage', [self.map.getId(), self.getId(), image]);
   });
   self.on('bounds_changed', function () {
     var bounds = self.get('bounds');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setBounds', [self.getId(), bounds]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setBounds', [self.map.getId(), self.getId(), bounds]);
   });
   self.on('opacity_changed', function () {
     var opacity = self.get('opacity');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setOpacity', [self.getId(), opacity]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setOpacity', [self.map.getId(), self.getId(), opacity]);
   });
   self.on('clickable_changed', function () {
     var clickable = self.get('clickable');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setClickable', [self.getId(), clickable]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setClickable', [self.map.getId(), self.getId(), clickable]);
   });
   self.on('bearing_changed', function () {
     var bearing = self.get('bearing');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setBearing', [self.getId(), bearing]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setBearing', [self.map.getId(), self.getId(), bearing]);
   });
   self.on('zIndex_changed', function () {
     var zIndex = self.get('zIndex');
-    self.exec.call(self, null, self.errorHandler, self.getPluginName(), 'setZIndex', [self.getId(), zIndex]);
+    self.exec.call(self, null, self.errorHandler, 'PluginGroundOverlay', 'setZIndex', [self.map.getId(), self.getId(), zIndex]);
   });
 
 };
@@ -136,7 +142,7 @@ GroundOverlay.prototype.remove = function (callback) {
         resolve.call(self);
       },
       reject.bind(self),
-      self.getPluginName(), 'remove', [self.getId()], {
+      'PluginGroundOverlay', 'remove', [self.map.getId(), self.getId()], {
         remove: true
       });
   };
