@@ -1,5 +1,4 @@
 
-
 var utils = require('cordova/utils'),
   cordova_exec = require('cordova/exec'),
   common = require('./Common'),
@@ -7,6 +6,7 @@ var utils = require('cordova/utils'),
   BaseClass = require('./BaseClass'),
   BaseArrayClass = require('./BaseArrayClass'),
   LatLng = require('./LatLng'),
+  LatLngBounds = require('./LatLngBounds'),
   MapTypeId = require('./MapTypeId'),
   event = require('./event'),
   VisibleRegion = require('./VisibleRegion'),
@@ -18,12 +18,23 @@ var utils = require('cordova/utils'),
   GroundOverlay = require('./GroundOverlay'),
   KmlOverlay = require('./KmlOverlay'),
   KmlLoader = require('./KmlLoader'),
+  DirectionsRenderer = require('./DirectionsRenderer'),
+  spherical = require('./spherical'),
+  encoding = require('./encoding'),
   MarkerCluster = require('./MarkerCluster');
 
 /**
  * Google Maps model.
  */
 var exec;
+var _protect = function(listener) {
+  Object.defineProperty(listener, '_protect', {
+    enumerable: false,
+    value: true,
+    writable: false
+  });
+  return listener;
+}
 var Map = function(__pgmId, _exec) {
   var self = this;
   exec = _exec;
@@ -48,6 +59,21 @@ var Map = function(__pgmId, _exec) {
   infoWindowLayer.style.overflow = 'visible';
   infoWindowLayer.style['z-index'] = 1;
 
+  var style = document.createElement('style');
+  style.setAttribute('type', 'text/css');
+  style.innerHTML = [
+    '.pgm-html-info-frame > * {',
+      'margin: 0;',
+      'padding: 0;',
+      'border: 0;',
+      'font-size: 100%;',
+      'font: inherit;',
+      'vertical-align: baseline;',
+      'color: black;',
+    '}'
+  ].join("\n");
+  infoWindowLayer.appendChild(style);
+
   Object.defineProperty(self, '_layers', {
     value: {
       info: infoWindowLayer
@@ -56,17 +82,17 @@ var Map = function(__pgmId, _exec) {
     writable: false
   });
 
-  self.on(event.MAP_CLICK, function() {
+  self.on(event.MAP_CLICK, _protect(function() {
     self.set('active_marker', undefined);
-  });
+  }));
 
-  self.on('active_marker_changed', function(prevMarker, newMarker) {
+  self.on('active_marker_changed', _protect(function(prevMarker, newMarker) {
     var newMarkerId = newMarker ? newMarker.getId() : null;
     if (prevMarker) {
       prevMarker.hideInfoWindow.call(prevMarker);
     }
-    self.exec.call(self, null, null, self.__pgmId, 'setActiveMarkerId', [newMarkerId]);
-  });
+    self.exec.call(self, null, null, 'PluginMap', 'setActiveMarkerId', [self.__pgmId, newMarkerId]);
+  }));
 };
 
 utils.extend(Map, Overlay);
@@ -77,11 +103,13 @@ utils.extend(Map, Overlay);
 Map.prototype.refreshLayout = function() {
   // Webkit redraw mandatory
   // http://stackoverflow.com/a/3485654/697856
-  document.body.style.display = 'inline-block';
-  document.body.offsetHeight;
-  document.body.style.display = '';
+  // document.body.style.display = 'inline-block';
+  // document.body.offsetHeight;
+  // document.body.style.display = '';
+  document.body.style.transform = 'rotateZ(0deg)';
 
-  this.exec.call(this, null, null, this.__pgmId, 'resizeMap', []);
+  var self = this;
+  self.exec.call(self, null, null, 'CordovaGoogleMaps', 'resizeMap', [self.__pgmId]);
 };
 
 Map.prototype.getMap = function(meta, div, options) {
@@ -97,12 +125,53 @@ Map.prototype.getMap = function(meta, div, options) {
     this.set('myLocationButton', options.controls.myLocationButton === true);
   }
 
-  if (options.preferences && options.preferences.gestureBounds) {
-    if (utils.isArray(options.preferences.gestureBounds) ||
-        options.preferences.gestureBounds.type === 'LatLngBounds') {
-      options.preferences.gestureBounds = common.convertToPositionArray(options.preferences.gestureBounds);
+  if (options.preferences && options.preferences.restriction) {
+
+    var bounds = new LatLngBounds();
+    if (utils.isArray(options.preferences.restriction)) {
+      options.preferences.restriction.forEach(function(ele) {
+        if (ele.lat && ele.lng) {
+          bounds.extend(ele);
+        }
+      });
+    } else if (options.preferences.restriction.type === 'LatLngBounds' ||
+      options.preferences.restriction.northeast && options.preferences.restriction.southwest) {
+      bounds.extend(options.preferences.restriction.southwest);
+      bounds.extend(options.preferences.restriction.northeast);
+    }
+
+    if (!bounds.southwest || !bounds.northeast) {
+      console.warn('(getMap) options.preferences.restriction is invalid.');
+      delete options.preferences.restriction;
+    } else {
+      var minZoom = !div ? 0 : spherical.computeBoundsZoom(bounds, div.offsetWidth, div.offsetHeight, 256);
+      var maxZoom = 23;
+      var prefMinZoom = 0;
+      var prefMaxZoom = 23;
+      if (options.preferences.zoom) {
+        if (options.preferences.zoom.minZoom) {
+          minZoom = options.preferences.zoom.minZoom;
+          prefMinZoom = minZoom;
+        }
+        if (options.preferences.zoom.maxZoom) {
+          maxZoom = options.preferences.zoom.maxZoom;
+          maxZoom = minZoom;
+        }
+      }
+      options.preferences.restriction = {
+        'south': bounds.southwest.lat,
+        'west': bounds.southwest.lng,
+        'north': bounds.northeast.lat,
+        'east': bounds.northeast.lng,
+        'prefMinZoom': prefMinZoom,
+        'prefMaxZoom': prefMaxZoom,
+        'minZoom': minZoom,
+        'maxZoom': maxZoom
+      };
+      self.set('restriction', options.preferences.restriction);
     }
   }
+
 
   if (!common.isDom(div)) {
     self.set('visible', false);
@@ -184,9 +253,6 @@ Map.prototype.getMap = function(meta, div, options) {
     _isReady: true
   }, function() {
 
-    //------------------------------------------------------------------------
-    // Clear background colors of map div parents after the map is created
-    //------------------------------------------------------------------------
     var div = self.get('div');
     if (common.isDom(div)) {
 
@@ -208,14 +274,6 @@ Map.prototype.getMap = function(meta, div, options) {
       div.insertBefore(self._layers.info, div.firstChild);
 
 
-      while (div.parentNode) {
-        div.style.backgroundColor = 'rgba(0,0,0,0) !important';
-
-        // Add _gmaps_cdv_ class
-        common.attachTransparentClass(div);
-
-        div = div.parentNode;
-      }
     }
     cordova.fireDocumentEvent('plugin_touch', {
       force: true
@@ -250,57 +308,83 @@ Map.prototype.getMap = function(meta, div, options) {
 Map.prototype.setOptions = function(options) {
   options = options || {};
 
+  var self = this;
+  var div = self.get('div');
   if (options.controls) {
-    var myLocation = this.get('myLocation');
-    if ('myLocation' in options.controls) {
-      myLocation = options.controls.myLocation === true;
-    }
-    var myLocationButton = this.get('myLocationButton');
-    if ('myLocationButton' in options.controls) {
-      myLocationButton = options.controls.myLocationButton === true;
-    }
-    this.set('myLocation', myLocation);
-    this.set('myLocationButton', myLocationButton);
-    if (myLocation === true || myLocation === false) {
-      options.controls.myLocation = myLocation;
-    }
-    if (myLocationButton === true || myLocationButton === false) {
-      options.controls.myLocationButton = myLocationButton;
-    }
+    self.set('myLocation', options.controls.myLocation === true);
+    self.set('myLocationButton', options.controls.myLocationButton === true);
   }
-  if (options.camera) {
-    if (options.camera.latLng) {
-      options.camera.target = options.camera.latLng;
-      delete options.camera.latLng;
-    }
-    this.set('camera', options.camera);
-    if (options.camera.target) {
-      this.set('camera_target', options.camera.target);
-    }
-    if (options.camera.bearing) {
-      this.set('camera_bearing', options.camera.bearing);
-    }
-    if (options.camera.zoom) {
-      this.set('camera_zoom', options.camera.zoom);
-    }
-    if (options.camera.tilt) {
-      this.set('camera_tilt', options.camera.tilt);
+
+  // if (options.camera && utils.isArray(options.camera.target)) {
+  //   var cameraBounds = new LatLngBounds();
+  //   options.camera.target.forEach(function(ele) {
+  //     if (ele.lat && ele.lng) {
+  //       cameraBounds.extend(ele);
+  //     }
+  //   });
+  //   options.camera.target = cameraBounds.getCenter();
+  //   options.camera.zoom = spherical.computeBoundsZoom(cameraBounds, div.offsetWidth, div.offsetHeight, 256);
+  // }
+  if (options.preferences) {
+    if (options.preferences.restriction) {
+
+      var bounds = new LatLngBounds();
+      if (utils.isArray(options.preferences.restriction)) {
+        options.preferences.restriction.forEach(function(ele) {
+          if (ele.lat && ele.lng) {
+            bounds.extend(ele);
+          }
+        });
+      } else if (options.preferences.restriction.type === 'LatLngBounds' ||
+        options.preferences.restriction.northeast && options.preferences.restriction.southwest) {
+        bounds.extend(options.preferences.restriction.southwest);
+        bounds.extend(options.preferences.restriction.northeast);
+      }
+
+      if (!bounds.southwest || !bounds.northeast) {
+        console.warn('(setOptions) options.preferences.restriction is invalid.');
+        delete options.preferences.restriction;
+      } else {
+        var minZoom = !div ? 0 : spherical.computeBoundsZoom(bounds, div.offsetWidth, div.offsetHeight, 256);
+        var maxZoom = 23;
+        var prefMinZoom = 0;
+        var prefMaxZoom = 23;
+        if (options.preferences.zoom) {
+          if (options.preferences.zoom.minZoom) {
+            minZoom = options.preferences.zoom.minZoom;
+            prefMinZoom = minZoom;
+          }
+          if (options.preferences.zoom.maxZoom) {
+            maxZoom = options.preferences.zoom.maxZoom;
+            maxZoom = minZoom;
+          }
+        }
+        options.preferences.restriction = {
+          'south': bounds.southwest.lat,
+          'west': bounds.southwest.lng,
+          'north': bounds.northeast.lat,
+          'east': bounds.northeast.lng,
+          'prefMinZoom': prefMinZoom,
+          'prefMaxZoom': prefMaxZoom,
+          'minZoom': minZoom,
+          'maxZoom': maxZoom
+        };
+        self.set('restriction', options.preferences.restriction);
+      }
+    } else {
+      self.set('restriction', undefined);
     }
   }
 
-  if (options.preferences && options.preferences.gestureBounds) {
-    if (utils.isArray(options.preferences.gestureBounds) ||
-        options.preferences.gestureBounds.type === 'LatLngBounds') {
-      options.preferences.gestureBounds = common.convertToPositionArray(options.preferences.gestureBounds);
-    }
-  }
 
   if (utils.isArray(options.styles)) {
     options.styles = JSON.stringify(options.styles);
   }
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setOptions', [options]);
-  return this;
+  return (new Promise(function(resolve, reject) {
+    self.exec.call(self, resolve, reject, 'PluginMap', 'setOptions', [self.__pgmId, options]);
+  }));
 };
+
 
 Map.prototype.getMyLocation = function(params, success_callback, error_callback) {
   return window.plugin.google.maps.LocationService.getMyLocation.call(this, params, success_callback, error_callback);
@@ -308,13 +392,16 @@ Map.prototype.getMyLocation = function(params, success_callback, error_callback)
 
 Map.prototype.setCameraTarget = function(latLng) {
   this.set('camera_target', latLng);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setCameraTarget', [latLng.lat, latLng.lng]);
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setCameraTarget', [this.__pgmId, {
+    'lat': latLng.lat,
+    'lng': latLng.lng
+  }]);
   return this;
 };
 
 Map.prototype.setCameraZoom = function(zoom) {
   this.set('camera_zoom', zoom);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setCameraZoom', [zoom], {
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setCameraZoom', [this.__pgmId, zoom], {
     sync: true
   });
   return this;
@@ -322,7 +409,10 @@ Map.prototype.setCameraZoom = function(zoom) {
 Map.prototype.panBy = function(x, y) {
   x = parseInt(x, 10);
   y = parseInt(y, 10);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'panBy', [x, y], {
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'panBy', [this.__pgmId, {
+    'x': x,
+    'y': y
+  }], {
     sync: true
   });
   return this;
@@ -370,7 +460,7 @@ Map.prototype.clear = function(callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.__pgmId, 'clear', [], {
+      'PluginMap', 'clear', [self.__pgmId], {
         sync: true
       });
   };
@@ -397,7 +487,7 @@ Map.prototype.setMapTypeId = function(mapTypeId) {
     return this.errorHandler('Invalid MapTypeId was specified.');
   }
   this.set('mapTypeId', mapTypeId);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setMapTypeId', [mapTypeId]);
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setMapTypeId', [this.__pgmId, mapTypeId]);
   return this;
 };
 
@@ -407,7 +497,7 @@ Map.prototype.setMapTypeId = function(mapTypeId) {
  */
 Map.prototype.setCameraTilt = function(tilt) {
   this.set('camera_tilt', tilt);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setCameraTilt', [tilt], {
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setCameraTilt', [this.__pgmId, tilt], {
     sync: true
   });
   return this;
@@ -419,7 +509,7 @@ Map.prototype.setCameraTilt = function(tilt) {
  */
 Map.prototype.setCameraBearing = function(bearing) {
   this.set('camera_bearing', bearing);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setCameraBearing', [bearing], {
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setCameraBearing', [this.__pgmId, bearing], {
     sync: true
   });
   return this;
@@ -491,7 +581,7 @@ Map.prototype.animateCamera = function(cameraPosition, callback) {
   //   cameraPosition.padding = 10;
   // }
 
-  if (utils.isArray(target) || target.type === 'LatLngBounds') {
+  if (utils.isArray(target) || target.type === 'LatLngBounds' || target.southwest && target.northeast) {
     target = common.convertToPositionArray(target);
     if (target.length === 0) {
       // skip if no point is specified
@@ -510,7 +600,9 @@ Map.prototype.animateCamera = function(cameraPosition, callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.__pgmId, 'animateCamera', [cameraPosition], {
+      'PluginMap',
+      'animateCamera',
+      [self.__pgmId, cameraPosition], {
         sync: true
       });
   };
@@ -561,7 +653,9 @@ Map.prototype.moveCamera = function(cameraPosition, callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.__pgmId, 'moveCamera', [cameraPosition], {
+      'PluginMap',
+      'moveCamera',
+      [self.__pgmId, cameraPosition], {
         sync: true
       });
   };
@@ -577,42 +671,52 @@ Map.prototype.setMyLocationButtonEnabled = function(enabled) {
   var self = this;
   enabled = common.parseBoolean(enabled);
   this.set('myLocationButton', enabled);
-  self.exec.call(self, null, this.errorHandler, this.__pgmId, 'setMyLocationEnabled', [{
-    myLocationButton: enabled,
-    myLocation: self.get('myLocation') === true
-  }], {
-    sync: true
-  });
+  self.exec.call(self, null, self.errorHandler,
+    'PluginMap',
+    'setMyLocationEnabled',
+    [
+      self.__pgmId,
+      {
+        myLocationButton: enabled,
+        myLocation: self.get('myLocation') === true
+      }], {
+      sync: true
+    });
   return this;
 };
 
 Map.prototype.setMyLocationEnabled = function(enabled) {
   var self = this;
   enabled = common.parseBoolean(enabled);
-  this.set('myLocation', enabled);
-  self.exec.call(self, null, this.errorHandler, this.__pgmId, 'setMyLocationEnabled', [{
-    myLocationButton: self.get('myLocationButton') === true,
-    myLocation: enabled
-  }], {
-    sync: true
-  });
+  self.set('myLocation', enabled);
+  self.exec.call(self, null, self.errorHandler,
+    'PluginMap',
+    'setMyLocationEnabled',
+    [
+      self.__pgmId,
+      {
+        myLocationButton: self.get('myLocationButton') === true,
+        myLocation: enabled
+      }], {
+        sync: true
+      });
   return this;
 };
 
 Map.prototype.setIndoorEnabled = function(enabled) {
   enabled = common.parseBoolean(enabled);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setIndoorEnabled', [enabled]);
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setIndoorEnabled', [this.__pgmId, enabled]);
   return this;
 };
 Map.prototype.setTrafficEnabled = function(enabled) {
   enabled = common.parseBoolean(enabled);
-  this.exec.call(this, null, this.errorHandler, this.__pgmId, 'setTrafficEnabled', [enabled]);
+  this.exec.call(this, null, this.errorHandler, 'PluginMap', 'setTrafficEnabled', [this.__pgmId, enabled]);
   return this;
 };
 Map.prototype.setCompassEnabled = function(enabled) {
   var self = this;
   enabled = common.parseBoolean(enabled);
-  self.exec.call(self, null, self.errorHandler, this.__pgmId, 'setCompassEnabled', [enabled]);
+  self.exec.call(self, null, self.errorHandler, 'PluginMap', 'setCompassEnabled', [this.__pgmId, enabled]);
   return this;
 };
 Map.prototype.getFocusedBuilding = function(callback) {
@@ -621,7 +725,7 @@ Map.prototype.getFocusedBuilding = function(callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.__pgmId, 'getFocusedBuilding', []);
+      'PluginMap', 'getFocusedBuilding', [self.__pgmId]);
   };
 
   if (typeof callback === 'function') {
@@ -638,7 +742,7 @@ Map.prototype.setVisible = function(isVisible) {
   var self = this;
   isVisible = common.parseBoolean(isVisible);
   self.set('visible', isVisible);
-  self.exec.call(self, null, self.errorHandler, this.__pgmId, 'setVisible', [isVisible]);
+  self.exec.call(self, null, self.errorHandler, 'PluginMap', 'setVisible', [self.__pgmId, isVisible]);
   return this;
 };
 
@@ -647,7 +751,7 @@ Map.prototype.setClickable = function(isClickable) {
   var self = this;
   isClickable = common.parseBoolean(isClickable);
   self.set('clickable', isClickable);
-  self.exec.call(self, null, self.errorHandler, this.__pgmId, 'setClickable', [isClickable]);
+  self.exec.call(self, null, self.errorHandler, 'PluginMap', 'setClickable', [self.__pgmId, isClickable]);
   return this;
 };
 Map.prototype.getClickable = function() {
@@ -661,7 +765,7 @@ Map.prototype.getClickable = function() {
 Map.prototype.setAllGesturesEnabled = function(enabled) {
   var self = this;
   enabled = common.parseBoolean(enabled);
-  self.exec.call(self, null, self.errorHandler, this.__pgmId, 'setAllGesturesEnabled', [enabled]);
+  self.exec.call(self, null, self.errorHandler, 'PluginMap', 'setAllGesturesEnabled', [self.__pgmId, enabled]);
   return this;
 };
 
@@ -680,7 +784,7 @@ Map.prototype.getCameraPosition = function() {
 Map.prototype.stopAnimation = function() {
   var self = this;
   if (self._isReady) {
-    cordova_exec(null, null, self.__pgmId, 'stopAnimation', []);
+    cordova_exec(null, null, 'PluginMap', 'stopAnimation', [self.__pgmId]);
   }
 };
 
@@ -697,6 +801,7 @@ Map.prototype.remove = function(callback) {
     writable: false
   });
   self.stopAnimation();
+  self.set('div', null);
 
   self.trigger('remove');
   // var div = self.get('div');
@@ -779,7 +884,7 @@ Map.prototype.toDataURL = function(params, callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.__pgmId, 'toDataURL', [params]);
+      'PluginMap', 'toDataURL', [self.__pgmId, params]);
   };
 
   if (typeof callback === 'function') {
@@ -803,14 +908,22 @@ Map.prototype.setDiv = function(div) {
   var self = this,
     args = [];
 
-  if (!common.isDom(div)) {
-    div = self.get('div');
-    if (common.isDom(div)) {
-      div.removeAttribute('__pluginMapId');
+  var prevDiv = self.get('div');
+  if (common.isDom(prevDiv)) {
+    prevDiv.__pluginMapId = undefined;
+  }
+
+  if (common.isDom(div)) {
+    var elemId = common.getPluginDomId(div);
+    args.push(elemId);
+
+    if (!div.__pluginMapId) {
+      Object.defineProperty(div, '__pluginMapId', {
+        enumerable: false,
+        writable: true,
+        value: self.__pgmId
+      });
     }
-    self.set('div', null);
-  } else {
-    div.setAttribute('__pluginMapId', self.__pgmId);
 
     // Insert the infoWindow layer
     if (self._layers.info.parentNode) {
@@ -835,39 +948,25 @@ Map.prototype.setDiv = function(div) {
     // div.offsetHeight;
     // div.style.display = '';
     document.body.style.transform = 'rotateZ(0deg)';
+    div.style.overflow = 'hidden';
+    div.style.position = 'relative';
 
-    self.set('div', div);
-
-    if (cordova.platform === 'browser') {
-      return;
-    }
-
-
-    positionCSS = common.getStyle(div, 'position');
-    if (!positionCSS || positionCSS === 'static') {
-      div.style.position = 'relative';
-    }
-    var elemId = common.getPluginDomId(div);
-    args.push(elemId);
-    while (div.parentNode) {
-      div.style.backgroundColor = 'rgba(0,0,0,0)';
-
-      // Add _gmaps_cdv_ class
-      common.attachTransparentClass(div);
-
-      div = div.parentNode;
-    }
   }
-  self.exec.call(self, function() {
-    cordova.fireDocumentEvent('plugin_touch', {
-      force: true,
-      action: 'setDiv'
+  self.set('div', div);
+  args.unshift(self.__pgmId);
+
+  return (new Promise(function(resolve) {
+    self.exec.call(self, function() {
+      cordova.fireDocumentEvent('plugin_touch', {
+        force: true,
+        action: 'setDiv'
+      });
+      self.refreshLayout();
+      resolve();
+    }, self.errorHandler, 'CordovaGoogleMaps', 'setDiv', args, {
+      sync: true
     });
-    self.refreshLayout();
-  }, self.errorHandler, self.__pgmId, 'setDiv', args, {
-    sync: true
-  });
-  return self;
+  }));
 };
 
 /**
@@ -908,7 +1007,12 @@ Map.prototype.fromLatLngToPoint = function(latLng, callback) {
       self.exec.call(self,
         resolve.bind(self),
         reject.bind(self),
-        self.__pgmId, 'fromLatLngToPoint', [latLng.lat, latLng.lng]);
+        'PluginMap',
+        'fromLatLngToPoint',
+        [self.__pgmId, {
+          'lat': latLng.lat,
+          'lng': latLng.lng
+        }]);
     };
 
     if (typeof callback === 'function') {
@@ -946,7 +1050,11 @@ Map.prototype.fromPointToLatLng = function(pixel, callback) {
           resolve.call(self, latLng);
         },
         reject.bind(self),
-        self.__pgmId, 'fromPointToLatLng', [pixel[0], pixel[1]]);
+        'PluginMap', 'fromPointToLatLng', [self.__pgmId,
+          {
+            'x': pixel[0],
+            'y': pixel[1]
+          }]);
     };
 
     if (typeof callback === 'function') {
@@ -1003,7 +1111,8 @@ Map.prototype.setPadding = function(p1, p2, p3, p4) {
     padding.left = padding.top;
     break;
   }
-  this.exec.call(this, null, self.errorHandler, this.__pgmId, 'setPadding', [padding]);
+  this.exec.call(this, null, self.errorHandler,
+    'PluginMap', 'setPadding', [this.__pgmId, padding]);
   return this;
 };
 
@@ -1068,6 +1177,26 @@ Map.prototype.addKmlOverlay = function(kmlOverlayOptions, callback) {
 };
 
 
+Map.prototype.addDirectionsRenderer = function(directionsRendererOptions) {
+  var self = this;
+  directionsRendererOptions = directionsRendererOptions || {};
+
+  if (directionsRendererOptions.directions) {
+
+    var renderer = new DirectionsRenderer(self, self.exec, directionsRendererOptions);
+    renderer.setRouteIndex(0);
+
+    return renderer;
+  } else {
+
+    if (typeof callback === 'function') {
+      throw new Error('directionsRendererOptions.directions is required.');
+    } else {
+      return Promise.reject('directionsRendererOptions.directions is required.');
+    }
+  }
+};
+
 //-------------
 // Ground overlay
 //-------------
@@ -1082,6 +1211,14 @@ Map.prototype.addGroundOverlay = function(groundOverlayOptions, callback) {
   groundOverlayOptions.zIndex = groundOverlayOptions.zIndex || 0;
   groundOverlayOptions.bounds = common.convertToPositionArray(groundOverlayOptions.bounds);
   groundOverlayOptions.noCaching = true;
+
+  if (groundOverlayOptions.url && groundOverlayOptions.url.indexOf("://") === -1) {
+
+    var link = document.createElement('a');
+    link.href = groundOverlayOptions.url;
+    groundOverlayOptions.url = link.protocol+'//'+link.host+link.pathname + link.search;
+    link = undefined;
+  }
 
   var groundOverlay = new GroundOverlay(self, groundOverlayOptions, exec);
   var groundOverlayId = groundOverlay.getId();
@@ -1098,7 +1235,7 @@ Map.prototype.addGroundOverlay = function(groundOverlayOptions, callback) {
     if (typeof callback === 'function') {
       callback.call(self, groundOverlay);
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['GroundOverlay', groundOverlayOptions, groundOverlay.hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'GroundOverlay', groundOverlayOptions, groundOverlay.hashCode]);
 
   return groundOverlay;
 };
@@ -1150,6 +1287,8 @@ Map.prototype.addTileOverlay = function(tilelayerOptions, callback) {
     debug: tilelayerOptions.debug
   };
 
+  var tileOverlayId = "tileoverlay_" + hashCode;
+  var link = document.createElement('a');
   var onNativeCallback = function(params) {
     var url = tilelayerOptions.getTile(params.x, params.y, params.zoom);
     if (!url || url === '(null)' || url === 'undefined' || url === 'null') {
@@ -1163,17 +1302,16 @@ Map.prototype.addTileOverlay = function(tilelayerOptions, callback) {
           link.href = finalUrl;
           finalUrl = link.protocol+'//'+link.host+link.pathname + link.search;
 
-          cordova_exec(null, self.errorHandler, self.__pgmId + '-tileoverlay', 'onGetTileUrlFromJS', [hashCode, params.key, finalUrl]);
+          cordova_exec(null, self.errorHandler, 'PluginTileOverlay', 'onGetTileUrlFromJS', [self.map.getId(), tileOverlayId, params.key, finalUrl]);
         })
         .catch(function() {
-          cordova_exec(null, self.errorHandler, self.__pgmId + '-tileoverlay', 'onGetTileUrlFromJS', [hashCode, params.key, '(null)']);
+          cordova_exec(null, self.errorHandler, 'PluginTileOverlay', 'onGetTileUrlFromJS', [self.map.getId(), tileOverlayId, params.key, '(null)']);
         });
     } else {
 
-      var link = document.createElement('a');
       link.href = url;
       url = link.protocol+'//'+link.host+link.pathname + link.search;
-      cordova_exec(null, self.errorHandler, self.__pgmId + '-tileoverlay', 'onGetTileUrlFromJS', [hashCode, params.key, url]);
+      cordova_exec(null, self.errorHandler, 'PluginTileOverlay', 'onGetTileUrlFromJS', [self.map.getId(), tileOverlayId, params.key, url]);
     }
   };
   document.addEventListener(self.__pgmId + '-' + hashCode + '-tileoverlay', onNativeCallback);
@@ -1185,10 +1323,11 @@ Map.prototype.addTileOverlay = function(tilelayerOptions, callback) {
     if (typeof callback === 'function') {
       callback.call(self, tileOverlay);
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['TileOverlay', options, hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'TileOverlay', options, hashCode]);
 
   return tileOverlay;
 };
+
 
 //-------------
 // Polygon
@@ -1196,6 +1335,9 @@ Map.prototype.addTileOverlay = function(tilelayerOptions, callback) {
 Map.prototype.addPolygon = function(polygonOptions, callback) {
   var self = this;
   polygonOptions.points = polygonOptions.points || [];
+  if (typeof polygonOptions.points === 'string') {
+    polygonOptions.points = encoding.decodePath(polygonOptions.points);
+  }
   var _orgs = polygonOptions.points;
   polygonOptions.points = common.convertToPositionArray(polygonOptions.points);
   polygonOptions.holes = polygonOptions.holes || [];
@@ -1243,7 +1385,7 @@ Map.prototype.addPolygon = function(polygonOptions, callback) {
     if (typeof callback === 'function') {
       callback.call(self, polygon);
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['Polygon', opts, polygon.hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'Polygon', opts, polygon.hashCode]);
 
   return polygon;
 };
@@ -1254,6 +1396,9 @@ Map.prototype.addPolygon = function(polygonOptions, callback) {
 Map.prototype.addPolyline = function(polylineOptions, callback) {
   var self = this;
   polylineOptions.points = polylineOptions.points || [];
+  if (typeof polylineOptions.points === 'string') {
+    polylineOptions.points = encoding.decodePath(polylineOptions.points);
+  }
   var _orgs = polylineOptions.points;
   polylineOptions.points = common.convertToPositionArray(polylineOptions.points);
   polylineOptions.color = common.HTMLColor2RGBA(polylineOptions.color || '#FF000080', 0.75);
@@ -1284,7 +1429,7 @@ Map.prototype.addPolyline = function(polylineOptions, callback) {
         callback.call(self, polyline);
       }
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['Polyline', opts, polyline.hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'Polyline', opts, polyline.hashCode]);
 
   return polyline;
 };
@@ -1320,7 +1465,7 @@ Map.prototype.addCircle = function(circleOptions, callback) {
     if (typeof callback === 'function') {
       callback.call(self, circle);
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['Circle', circleOptions, circle.hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'Circle', circleOptions, circle.hashCode]);
 
   return circle;
 };
@@ -1392,7 +1537,7 @@ Map.prototype.addMarker = function(markerOptions, callback) {
         callback.call(self, marker);
       }
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['Marker', markerOptions, marker.hashCode]);
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId, 'Marker', markerOptions, marker.hashCode]);
 
   return marker;
 };
@@ -1468,7 +1613,8 @@ Map.prototype.addMarkerCluster = function(markerClusterOptions, callback) {
     if (typeof callback === 'function') {
       callback.call(self, markerCluster);
     }
-  }, self.errorHandler, self.__pgmId, 'loadPlugin', ['MarkerCluster', {
+  }, self.errorHandler, 'PluginMap', 'loadPlugin', [self.__pgmId,
+    'MarkerCluster', {
     'positionList': positionList,
     'debug': markerClusterOptions.debug === true
   }, markerCluster.hashCode]);
