@@ -163,7 +163,9 @@ var MarkerCluster = function (map, markerClusterOptions, _exec) {
 
         link = document.createElement('a');
         link.href = markerOptions.icon;
-        markerOptions.icon = link.protocol + '//' + link.host + link.pathname + link.search;
+        markerOptions.icon = {
+          'url': link.protocol + '//' + link.host + link.pathname + link.search
+        };
         link = undefined;
       }
     } else if (typeof markerOptions.icon === 'object' && typeof markerOptions.icon.url === 'string') {
@@ -193,7 +195,7 @@ var MarkerCluster = function (map, markerClusterOptions, _exec) {
     });
     return marker;
   };
-  
+
   self.addMarkers = function (markers) {
     var results = [];
     if (utils.isArray(markers) || Array.isArray(markers)) {
@@ -263,8 +265,18 @@ MarkerCluster.prototype.onClusterClicked = function (cluster) {
   var zoomLevel = computeZoom(cluster.getBounds(), self.map.getDiv());
   zoomLevel += zoomLevel === self.map.get('camera_zoom') ? 1 : 0;
   self.map.animateCamera({
-    target: cluster.getBounds().getCenter(),
-    zoom: zoomLevel,
+    target: [
+      bounds.southwest,
+      {
+        lat: bounds.northeast.lat,
+        lng: bounds.southwest.lng
+      },
+      bounds.northeast,
+      {
+        lat: bounds.southwest.lat,
+        lng: bounds.northeast.lng
+      }
+    ],
     duration: 500
   }, function () {
     if (self.boundsDraw) {
@@ -361,7 +373,7 @@ MarkerCluster.prototype.remove = function (callback) {
     self.exec.call(self,
       resolve.bind(self),
       reject.bind(self),
-      self.getPluginName(), 'remove', [self.getId()], {
+      'PluginMarkerCluster', 'remove', [self.map.getId(), self.getId()], {
         sync: true,
         remove: true
       });
@@ -427,7 +439,7 @@ Object.defineProperty(MarkerCluster.prototype, '_removeMarkerById', {
     marker.destroy();
     delete self._markerMap[markerId];
     if (isAdded) {
-      self.exec.call(self, null, null, self.getPluginName(), 'redrawClusters', [self.getId(), {
+      self.exec.call(self, null, null, 'PluginMarkerCluster', 'redrawClusters', [self.map.getId(), self.getId(), {
         'delete': [markerId]
       }], {
         sync: true
@@ -1120,19 +1132,26 @@ Object.defineProperty(MarkerCluster.prototype, '_redraw', {
         if (typeof marker.get('icon') === 'string') {
           marker.set('icon', {
             'url': marker.get('icon'),
-            'size': size,
+            'size': {
+              'width': Math.floor(size.width),
+              'height': Math.floor(size.height)
+            },
             'anchor': [size.width / 2, size.height]
           }, true);
         } else {
           var icon = marker.get('icon') || {};
           icon.size = icon.size || size;
+          icon.size = {
+            'width': Math.floor(icon.size.width),
+            'height': Math.floor(icon.size.height)
+          };
           icon.anchor = icon.anchor || [size.width / 2, size.height];
           self._markerMap[markerId].set('icon', icon, true);
         }
         marker.set('infoWindowAnchor', marker.get('infoWindowAnchor') || [marker.get('icon').size.width / 2, 0], true);
       });
       self.trigger('nextTask');
-    }, self.errorHandler, self.getPluginName(), 'redrawClusters', [self.getId(), {
+    }, self.errorHandler, 'PluginMarkerCluster', 'redrawClusters', [self.map.getId(), self.getId(), {
       'resolution': resolution,
       'new_or_update': new_or_update_clusters,
       'delete': delete_clusters
@@ -1178,7 +1197,7 @@ MarkerCluster.prototype._createMarker = function (markerOpts) {
   var self = this;
   var markerId = self.getId() + '-marker_' + Math.floor(Date.now() * Math.random());
   var marker = new Marker(self.getMap(), markerOpts, exec, {
-    className: 'MarkerCluster',
+    type: 'MarkerCluster',
     __pgmId: markerId
   });
   marker._privateInitialize(markerOpts);
