@@ -1,12 +1,12 @@
 
 
 if (!window.Promise) {
-  window.Promise = require('cordova-plugin-googlemaps.Promise');
+  window.Promise = require('cordova-plugin-googlemaps-beta.Promise');
 }
 
-var common = require('cordova-plugin-googlemaps.Common'),
-  Map = require('cordova-plugin-googlemaps.Map'),
-  StreetViewPanorama = require('cordova-plugin-googlemaps.StreetViewPanorama');
+var common = require('cordova-plugin-googlemaps-beta.Common'),
+  Map = require('cordova-plugin-googlemaps-beta.Map'),
+  StreetViewPanorama = require('cordova-plugin-googlemaps-beta.StreetViewPanorama');
 
 
 function nativeCallback(params) {
@@ -29,8 +29,8 @@ function CordovaGoogleMaps(execCmd) {
   self.MAP_CNT = 0;
 
   var removeMapDiv = function(node) {
-    if (node.hasAttribute('__pluginmapid') && !node.parentNode) {
-      var mapId = node.getAttribute('__pluginmapid');
+    if (node.__pluginMapId !== undefined && !node.parentNode) {
+      var mapId = node.__pluginMapId;
       var map = self.MAPS[mapId];
       if (map) {
         map.remove();
@@ -39,9 +39,12 @@ function CordovaGoogleMaps(execCmd) {
     } else {
       var childNodes = Array.prototype.slice.call(node.childNodes);
       childNodes.forEach(function(child) {
-        if (child.outerHTML && child.outerHTML.indexOf('__pluginmapid') > -1) {
+        if (child.outerHTML && child.__pluginMapId) {
           removeMapDiv(child);
         }
+        // if (child.outerHTML && child.outerHTML.indexOf('__pluginmapid') > -1) {
+        //   removeMapDiv(child);
+        // }
       });
     }
   };
@@ -56,9 +59,12 @@ function CordovaGoogleMaps(execCmd) {
         if (record.removedNodes.length > 0) {
           record.removeNodes = Array.prototype.slice.call(record.removedNodes, 0);
           record.removeNodes.forEach(function(node) {
-            if (node.outerHTML && node.outerHTML.indexOf('__pluginmapid') > -1) {
+            if (node.outerHTML && node.__pluginMapId) {
               removeMapDiv(node);
             }
+            // if (node.outerHTML && node.outerHTML.indexOf('__pluginmapid') > -1) {
+            //   removeMapDiv(node);
+            // }
           });
         }
       });
@@ -89,7 +95,7 @@ CordovaGoogleMaps.prototype.getMap = function(div, mapOptions) {
     mapId;
 
   if (common.isDom(div)) {
-    mapId = div.getAttribute('__pluginMapId');
+    mapId = div.__pluginMapId;
 
     // Wow, the app specifies the map div that has already another map,
     // but the app try to create new map.
@@ -179,10 +185,19 @@ CordovaGoogleMaps.prototype._remove = function(mapId) {
 
     var div = map.getDiv();
     if (!div) {
-      div = document.querySelector('[__pluginMapId="' + mapId + '"]');
+      var eles = Array.from(document.querySelectorAll('*'));
+      eles = eles.filter(function(e) {
+        return e.__pluginMapId === mapId;
+      });
+      if (eles.length === 1) {
+        div = eles[0];
+      }
     }
     if (div) {
-      div.removeAttribute('__pluginMapId');
+      Object.defineProperty(div, '__pluginMapId', {
+        enumerable: false,
+        value: undefined
+      });
     }
 
     self.MAPS[mapId].destroy();
@@ -215,7 +230,10 @@ function postPanoramaInit(panorama, div, options) {
   // If the mapDiv is specified,
   // the native side needs to know the map div position
   // before creating the map view.
-  div.setAttribute('__pluginMapId', mapId);
+  Object.defineProperty(div, '__pluginMapId', {
+    enumerable: false,
+    value: mapId
+  });
 
   if (div.offsetWidth < 100 || div.offsetHeight < 100) {
     console.error('[GoogleMaps] Minimum container dimention is 100x100 in pixels.', div);
@@ -229,7 +247,10 @@ function postPanoramaInit(panorama, div, options) {
   // If the mapDiv is specified,
   // the native side needs to know the map div position
   // before creating the map view.
-  div.setAttribute('__pluginMapId', mapId);
+  Object.defineProperty(div, '__pluginMapId', {
+    enumerable: false,
+    value: mapId
+  });
 
   panorama.getPanorama.apply(panorama, args);
 }
@@ -254,7 +275,10 @@ function postMapInit(map, div, options) {
     // If the mapDiv is specified,
     // the native side needs to know the map div position
     // before creating the map view.
-    div.setAttribute('__pluginMapId', mapId);
+    Object.defineProperty(div, '__pluginMapId', {
+      enumerable: false,
+      value: mapId
+    });
 
     args.push({
       __pgmId: mapId,
